Reduce build time by moving headers inline implementations to src/
