+++ CMakeLists.txt
@@ -7,6 +7,7 @@ project(CipherTrader VERSION 1.0.0 LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF) # TODO
+set(CMAKE_UNITY_BUILD ON)
 
 # Compiler configuration
 if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
