+++ include/DB.hpp
@@ -1,49 +1,11 @@
 #ifndef CIPHER_DB_HPP
 #define CIPHER_DB_HPP
 
-#include <any>
-#include <atomic>
-#include <chrono>
-#include <csignal>
-#include <exception>
-#include <functional>
-#include <future>
-#include <limits>
-#include <memory>
-#include <mutex>
-#include <optional>
-#include <queue>
-#include <stdexcept>
-#include <string>
-#include <thread>
-#include <unordered_map>
-#include <vector>
-#include <blaze/Math.h>
+#include "Precompiled.hpp"
 
-#include "Config.hpp"
 #include "DynamicArray.hpp"
 #include "Enum.hpp"
-#include "Helper.hpp"
 #include "Logger.hpp"
-
-#include <blaze/Math.h>
-#include <boost/uuid.hpp>
-#include <boost/uuid/uuid.hpp>
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
-#include <nlohmann/json.hpp>
-#include <sqlpp11/char_sequence.h>
-#include <sqlpp11/data_types.h>
-#include <sqlpp11/data_types/blob/data_type.h>
-#include <sqlpp11/insert.h>
-#include <sqlpp11/null.h>
-#include <sqlpp11/postgresql/connection.h>
-#include <sqlpp11/postgresql/connection_config.h>
-#include <sqlpp11/postgresql/postgresql.h>
-#include <sqlpp11/sqlpp11.h>
-#include <sqlpp11/table.h>
-#include <sqlpp11/transaction.h>
-#include <sqlpp11/update.h>
 
 namespace ct
 {
@@ -56,60 +18,24 @@ class DatabaseShutdownManager
     using ShutdownHook           = std::function< void() >;
     using ShutdownCompletionHook = std::function< void() >;
 
-    static DatabaseShutdownManager& getInstance()
-    {
-        static DatabaseShutdownManager instance;
-        return instance;
-    }
+    static DatabaseShutdownManager& getInstance();
 
     // Register a hook to be called during shutdown
-    void registerShutdownHook(ShutdownHook hook)
-    {
-        std::lock_guard< std::mutex > lock(hooksMutex_);
-        shutdownHooks_.push_back(std::move(hook));
-    }
+    void registerShutdownHook(ShutdownHook hook);
 
     // Register a hook to be called after shutdown is complete
-    void registerCompletionHook(ShutdownCompletionHook hook)
-    {
-        std::lock_guard< std::mutex > lock(completionHooksMutex_);
-        completionHooks_.push_back(std::move(hook));
-    }
+    void registerCompletionHook(ShutdownCompletionHook hook);
 
     // Initialize signal handlers
-    void initSignalHandlers()
-    {
-        std::signal(SIGINT, handleSignal);
-        std::signal(SIGTERM, handleSignal);
-
-        // Ignore SIGPIPE to prevent crashes on broken connections
-        std::signal(SIGPIPE, SIG_IGN);
-    }
+    void initSignalHandlers();
 
     // Check if shutdown is in progress
-    bool isShuttingDown() const { return shuttingDown_.load(std::memory_order_acquire); }
+    bool isShuttingDown() const;
 
     // Wait for shutdown to complete
-    void waitForShutdown()
-    {
-        if (shutdownFuture_.valid())
-        {
-            shutdownFuture_.wait();
-        }
-    }
+    void waitForShutdown();
 
-    void shutdown()
-    {
-        // Only start shutdown once
-        bool expected = false;
-        if (!shuttingDown_.compare_exchange_strong(expected, true, std::memory_order_acq_rel))
-        {
-            return;
-        }
-
-        // Launch shutdown process asynchronously
-        shutdownFuture_ = std::async(std::launch::async, [this] { performShutdown(); });
-    }
+    void shutdown();
 
    private:
     DatabaseShutdownManager() : shuttingDown_(false) {}
@@ -118,7 +44,7 @@ class DatabaseShutdownManager
     DatabaseShutdownManager(const DatabaseShutdownManager&)            = delete;
     DatabaseShutdownManager& operator=(const DatabaseShutdownManager&) = delete;
 
-    static void handleSignal(int signal) { getInstance().shutdown(); }
+    static void handleSignal(int signal);
 
     void performShutdown();
 
@@ -135,11 +61,7 @@ class ConnectionPool
 {
    public:
     // Get singleton instance
-    static ConnectionPool& getInstance()
-    {
-        static ConnectionPool instance;
-        return instance;
-    }
+    static ConnectionPool& getInstance();
 
     // Initialize the connection pool
     void init(const std::string& host,
@@ -147,116 +69,18 @@ class ConnectionPool
               const std::string& username,
               const std::string& password,
               unsigned int port = 5432,
-              size_t poolSize   = 10) // Default pool size
-    {
-        std::lock_guard< std::mutex > lock(mutex_);
-
-        // Store connection parameters for creating new connections
-        host_     = host;
-        dbname_   = dbname;
-        username_ = username;
-        password_ = password;
-        port_     = port;
-
-        // Initialize the pool with connections
-        for (size_t i = 0; i < poolSize; ++i)
-        {
-            createNewConnection();
-        }
-    }
+              size_t poolSize   = 10); // Default pool size
 
     // Get a connection from the pool (or create one if pool is empty)
-    std::shared_ptr< sqlpp::postgresql::connection > getConnection()
-    {
-        std::unique_lock< std::mutex > lock(mutex_);
-
-        if (!initialized_)
-        {
-            throw std::runtime_error("Connection pool not initialized");
-        }
-
-        // Check if we're shutting down
-        if (DatabaseShutdownManager::getInstance().isShuttingDown())
-        {
-            throw std::runtime_error("Database is shutting down");
-        }
-
-        // Wait for a connection to become available
-        while (availableConnections_.empty())
-        {
-            // If we've reached max connections, wait for one to be returned
-            if (activeConnections_ >= maxConnections_)
-            {
-                // TODO: Read timeout from config.
-                if (!connectionAvailable_.wait_for(
-                        lock, std::chrono::seconds(5), [this] { return !availableConnections_.empty(); }))
-                {
-                    throw std::runtime_error("Timeout waiting for connection");
-                }
-            }
-            else
-            {
-                // Create a new connection if below max limit
-                createNewConnection();
-            }
-        }
-
-        // Get a connection from the pool
-        auto conn = std::move(availableConnections_.front());
-        availableConnections_.pop();
-
-        // Save the raw pointer before transferring ownership
-        sqlpp::postgresql::connection* rawConn = conn.get();
-
-        // IMPORTANT: Add to managed connections before creating the shared_ptr
-        managedConnections_.push_back(std::move(conn));
-        activeConnections_++;
-
-        // Create a wrapper that returns the connection to the pool when it's destroyed
-        return std::shared_ptr< sqlpp::postgresql::connection >(
-            rawConn, [this](sqlpp::postgresql::connection* conn) { this->returnConnection(conn); });
-    }
+    std::shared_ptr< sqlpp::postgresql::connection > getConnection();
 
     // Get a connection from the pool with health check
-    std::shared_ptr< sqlpp::postgresql::connection > getConnectionWithHealthCheck()
-    {
-        auto conn = getConnection();
-
-        // Test if the connection is still valid
-        try
-        {
-            conn->execute("SELECT 1");
-        }
-        catch (const std::exception& e)
-        {
-            // Connection is dead, get a new one
-            std::cerr << "Detected dead connection, getting a new one: " << e.what() << std::endl;
-
-            // This will release the dead connection and get a new one
-            // The connection pool's returnConnection will handle creating a replacement
-            conn = getConnection();
-        }
-
-        return conn;
-    }
+    std::shared_ptr< sqlpp::postgresql::connection > getConnectionWithHealthCheck();
 
     // Set maximum number of connections
-    void setMaxConnections(size_t maxConnections)
-    {
-        std::lock_guard< std::mutex > lock(mutex_);
-        maxConnections_ = maxConnections;
-    }
+    void setMaxConnections(size_t maxConnections);
 
-    void waitForConnectionsToClose()
-    {
-        std::unique_lock< std::mutex > lock(mutex_);
-
-        // Wait until all connections are returned
-        while (activeConnections_ > 0)
-        {
-            connectionReturned_.wait(lock);
-        }
-    }
+    void waitForConnectionsToClose();
 
     // Delete copy constructor and assignment operator
     ConnectionPool(const ConnectionPool&)            = delete;
@@ -264,87 +88,13 @@ class ConnectionPool
 
    private:
     // Private constructor for singleton
-    ConnectionPool() : activeConnections_(0), maxConnections_(20), initialized_(false) {}
+    ConnectionPool();
 
     // Return a connection to the pool
-    void returnConnection(sqlpp::postgresql::connection* conn)
-    {
-        std::lock_guard< std::mutex > lock(mutex_);
-
-        // Check if connection is still valid before trying to clean it up
-        bool connectionValid = true;
-        try
-        {
-            // A simple ping to test if connection is alive
-            conn->execute("SELECT 1");
-        }
-        catch (const std::exception& e)
-        {
-            // Connection is dead, mark it as invalid
-            connectionValid = false;
-            std::cerr << "Detected dead connection during return: " << e.what() << std::endl;
-        }
-
-        if (connectionValid)
-        {
-            // Only try to clean up if connection is still valid
-            try
-            {
-                conn->execute("DEALLOCATE ALL");
-            }
-            catch (const std::exception& e)
-            {
-                // Log the error but continue
-                std::cerr << "Error cleaning connection during return: " << e.what() << std::endl;
-                connectionValid = false;
-            }
-        }
-
-        // Find the connection in our managed connections
-        auto it = std::find_if(managedConnections_.begin(),
-                               managedConnections_.end(),
-                               [conn](const auto& managed) { return managed.get() == conn; });
-
-        if (it != managedConnections_.end())
-        {
-            if (!DatabaseShutdownManager::getInstance().isShuttingDown())
-            {
-                if (connectionValid)
-                {
-                    // Return to the pool only if valid
-                    availableConnections_.push(std::move(*it));
-                }
-                else
-                {
-                    // Connection is invalid, create a new one to replace it
-                    createNewConnection();
-                }
-            }
-
-            managedConnections_.erase(it);
-            activeConnections_--;
-
-            // Notify both waiting threads and shutdown manager
-            connectionAvailable_.notify_one();
-            connectionReturned_.notify_all();
-        }
-    }
+    void returnConnection(sqlpp::postgresql::connection* conn);
 
     // Create a new database connection
-    void createNewConnection()
-    {
-        auto config      = std::make_shared< sqlpp::postgresql::connection_config >();
-        config->debug    = false; // TODO: accept as arg
-        config->host     = host_;
-        config->dbname   = dbname_;
-        config->user     = username_;
-        config->password = password_;
-        config->port     = port_;
-
-        auto conn = std::make_unique< sqlpp::postgresql::connection >(config);
-        availableConnections_.push(std::move(conn));
-        initialized_ = true;
-    }
+    void createNewConnection();
 
     // Connection pool members
     std::mutex mutex_;
@@ -369,54 +119,21 @@ class Database
 {
    public:
     // Get singleton instance
-    static Database& getInstance()
-    {
-        static Database instance;
-        return instance;
-    }
+    static Database& getInstance();
 
     // Initialize the database with connection parameters
     void init(const std::string& host,
               const std::string& dbname,
               const std::string& username,
               const std::string& password,
-              unsigned int port = 5432)
-    {
-        std::lock_guard< std::mutex > lock(mutex_);
+              unsigned int port = 5432);
 
-        // Initialize connection pool
-        ConnectionPool::getInstance().init(host, dbname, username, password, port);
+    // Get a connection from the pool
+    // sqlpp::postgresql::connection& getConnection();
 
-        // Initialize shutdown manager
-        auto& shutdownManager = DatabaseShutdownManager::getInstance();
-        shutdownManager.initSignalHandlers();
+    std::shared_ptr< sqlpp::postgresql::connection > getConnection();
 
-        // Register shutdown hooks
-        shutdownManager.registerShutdownHook([] { std::cout << "Database shutdown initiated..." << std::endl; });
-
-        shutdownManager.registerCompletionHook([] { std::cout << "Database shutdown completed." << std::endl; });
-    }
-
-    // // Get a connection from the pool
-    // sqlpp::postgresql::connection& getConnection()
-    // {
-    // // Get connection from pool and cache it in thread_local storage
-    // thread_local std::shared_ptr< sqlpp::postgresql::connection > conn =
-    //     ConnectionPool::getInstance().getConnection();
-
-    // return *conn;
-    // }
-
-    std::shared_ptr< sqlpp::postgresql::connection > getConnection()
-    {
-        return ConnectionPool::getInstance().getConnection();
-    }
-
-    void shutdown()
-    {
-        DatabaseShutdownManager::getInstance().shutdown();
-        DatabaseShutdownManager::getInstance().waitForShutdown();
-    }
+    void shutdown();
 
     // Delete copy constructor and assignment operator
     Database(const Database&)            = delete;
@@ -434,139 +151,29 @@ class TransactionManager
 {
    public:
     // Start a new transaction
-    static std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > startTransaction()
-    {
-        auto db = db::Database::getInstance().getConnection();
-        return std::make_shared< sqlpp::transaction_t< sqlpp::postgresql::connection > >(start_transaction(*db));
-    }
+    static std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > startTransaction();
 
     // Commit the transaction
-    static bool commitTransaction(std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx)
-    {
-        if (!tx)
-        {
-            logger::LOG.error("Cannot commit null transaction");
-            return false;
-        }
-
-        try
-        {
-            tx->commit();
-            return true;
-        }
-        catch (const std::exception& e)
-        {
-            std::ostringstream oss;
-            oss << "Error committing transaction: " << e.what();
-            logger::LOG.error(oss.str());
-            return false;
-        }
-    }
+    static bool commitTransaction(std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx);
 
     // Roll back the transaction
-    static bool rollbackTransaction(std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx)
-    {
-        if (!tx)
-        {
-            std::ostringstream oss;
-            oss << "Cannot rollback null transaction";
-            logger::LOG.error(oss.str());
-            return false;
-        }
-
-        try
-        {
-            tx->rollback();
-            return true;
-        }
-        catch (const std::exception& e)
-        {
-            std::ostringstream oss;
-            oss << "Error rolling back transaction: " << e.what();
-            logger::LOG.error(oss.str());
-            return false;
-        }
-    }
+    static bool rollbackTransaction(std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx);
 };
 
 // RAII-style transaction guard
 class TransactionGuard
 {
    public:
-    TransactionGuard() : committed_(false)
-    {
-        // Check for shutdown before starting transaction
-        if (DatabaseShutdownManager::getInstance().isShuttingDown())
-        {
-            throw std::runtime_error("Cannot start new transaction during shutdown");
-        }
+    TransactionGuard();
 
-        conn_ = ConnectionPool::getInstance().getConnection();
+    ~TransactionGuard();
 
-        if (conn_)
-        {
-            tx_ = std::make_shared< sqlpp::transaction_t< sqlpp::postgresql::connection > >(start_transaction(*conn_));
-        }
-    }
+    bool commit();
 
-    ~TransactionGuard()
-    {
-        if (tx_ && !committed_)
-        {
-            try
-            {
-                tx_->rollback();
-            }
-            catch (const std::exception& e)
-            {
-                std::ostringstream oss;
-                oss << "Error during auto-rollback: " << e.what();
-                logger::LOG.error(oss.str());
-            }
-        }
-    }
-
-    bool commit()
-    {
-        if (!tx_)
-            return false;
-
-        try
-        {
-            tx_->commit();
-            committed_ = true;
-            return true;
-        }
-        catch (const std::exception& e)
-        {
-            std::ostringstream oss;
-            oss << "Error during commit: " << e.what();
-            logger::LOG.error(oss.str());
-            return false;
-        }
-    }
-
-    bool rollback()
-    {
-        if (!tx_)
-            return false;
-
-        try
-        {
-            tx_->rollback();
-            return true;
-        }
-        catch (const std::exception& e)
-        {
-            std::ostringstream oss;
-            oss << "Error during rollback: " << e.what();
-            logger::LOG.error(oss.str());
-            return false;
-        }
-    }
+    bool rollback();
 
     // Get the connection associated with this transaction
-    std::shared_ptr< sqlpp::postgresql::connection > getConnection() { return conn_; }
+    std::shared_ptr< sqlpp::postgresql::connection > getConnection();
 
     // Delete copy constructor/assignment
     TransactionGuard(const TransactionGuard&)            = delete;
@@ -579,58 +186,17 @@ class TransactionGuard
 };
 
 template < typename Func >
-auto executeWithRetry(Func&& operation, int maxRetries = 3) -> decltype(operation())
-{
-    int retries = 0;
-    while (true)
-    {
-        try
-        {
-            return operation();
-        }
-        catch (const std::exception& e)
-        {
-            if (++retries > maxRetries)
-            {
-                throw; // Re-throw after max retries
-            }
-
-            std::ostringstream oss;
-            oss << "Operation failed, retrying (" << retries << "/" << maxRetries << "): " << e.what();
-            logger::LOG.error(oss.str());
-
-            // Exponential backoff
-            std::this_thread::sleep_for(std::chrono::milliseconds(100 * retries));
-        }
-    }
-}
+auto executeWithRetry(Func&& operation, int maxRetries = 3) -> decltype(operation());
 
 // RAII class for handling connection state
 class ConnectionStateGuard
 {
    public:
-    explicit ConnectionStateGuard(sqlpp::postgresql::connection& conn) : conn_(conn), needs_reset_(false) {}
+    explicit ConnectionStateGuard(sqlpp::postgresql::connection& conn);
 
-    ~ConnectionStateGuard()
-    {
-        if (needs_reset_)
-        {
-            try
-            {
-                // Deallocate all prepared statements
-                conn_.execute("DEALLOCATE ALL");
+    ~ConnectionStateGuard();
 
-                // Execute a harmless query to reset the connection state
-                conn_.execute("SELECT 1");
-            }
-            catch (const std::exception& e)
-            {
-                // Nothing we can do in the destructor
-            }
-        }
-    }
-
-    void markForReset() { needs_reset_ = true; }
+    void markForReset();
 
    private:
     sqlpp::postgresql::connection& conn_;
@@ -640,224 +206,21 @@ class ConnectionStateGuard
 // Generic findById implementation
 template < typename ModelType >
 std::optional< ModelType > findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
-                                    const boost::uuids::uuid& id)
-{
-    // Use the provided connection if available, otherwise get the default connection
-    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
-    const auto& t = ModelType::table();
-
-    // Create state guard for this connection
-    ConnectionStateGuard stateGuard(conn);
-
-    try
-    {
-        // Convert UUID to string
-        std::string uuid_str = boost::uuids::to_string(id);
-
-        // Prepare statement
-        auto prep      = conn.prepare(select(all_of(t)).from(t).where(t.id == parameter(t.id)));
-        prep.params.id = uuid_str;
-
-        // Execute
-        auto result = conn(prep);
-
-        if (result.empty())
-        {
-            return std::nullopt;
-        }
-
-        // Create and populate a new model instance
-        const auto& row = *result.begin();
-        return ModelType::fromRow(row);
-    }
-    catch (const std::exception& e)
-    {
-        std::ostringstream oss;
-        oss << "Error finding " << ModelType::modelName() << " by ID: " << e.what();
-        logger::LOG.error(oss.str());
-
-        // Mark the connection for reset
-        stateGuard.markForReset();
-
-        throw;
-    }
-}
+                                    const boost::uuids::uuid& id);
 
 // Generic findByFilter implementation
 template < typename ModelType, typename FilterType >
 std::optional< std::vector< ModelType > > findByFilter(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
-                                                       const FilterType& filter)
-{
-    // Use the provided connection if available, otherwise get the default connection
-    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
-    const auto& t = ModelType::table();
-
-    // Create state guard for this connection
-    ConnectionStateGuard stateGuard(conn);
-
-    try
-    {
-        // Build dynamic query
-        auto query = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
-
-        // Apply filter conditions
-        filter.applyToQuery(query, t);
-
-        // Execute query
-        auto rows = conn(query);
-
-        // Process results
-        std::vector< ModelType > results;
-        for (const auto& row : rows)
-        {
-            try
-            {
-                results.push_back(std::move(ModelType::fromRow(row)));
-            }
-            catch (const std::exception& e)
-            {
-                std::ostringstream oss;
-                oss << "Error processing row: " << e.what();
-                logger::LOG.error(oss.str());
-
-                // Mark the connection for reset
-                stateGuard.markForReset();
-
-                throw;
-            }
-        }
-
-        return results;
-    }
-    catch (const std::exception& e)
-    {
-        std::ostringstream oss;
-        oss << "Error in findByFilter for " << ModelType::modelName() << ": " << e.what();
-        logger::LOG.error(oss.str());
-
-        // Mark the connection for reset
-        stateGuard.markForReset();
-
-        throw;
-    }
-}
+                                                       const FilterType& filter);
 
 // Generic save implementation
 template < typename ModelType >
-void save(ModelType& model, std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const bool update_on_conflict)
-{
-    // Use the provided connection if available, otherwise get the default connection
-    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
-    const auto& t = ModelType::table();
-
-    // Create state guard for this connection
-    ConnectionStateGuard stateGuard(conn);
-
-    try
-    {
-        auto select_stmt = model.prepareSelectStatementForConflictCheck(t, conn);
-        auto sprep       = conn.prepare(select_stmt);
-        auto rows        = conn(sprep);
-
-        std::vector< ModelType > retrieved;
-
-        if (!rows.empty())
-        {
-            for (const auto& row : rows)
-            {
-                try
-                {
-                    retrieved.push_back(std::move(ModelType::fromRow(row)));
-                }
-                catch (const std::exception& e)
-                {
-                    std::ostringstream oss;
-                    oss << "Error processing row: " << e.what();
-                    logger::LOG.error(oss.str());
-
-                    // Mark the connection for reset
-                    stateGuard.markForReset();
-
-                    throw;
-                }
-            }
-        }
-
-        // TODO: Support batch update?
-        if (retrieved.size() > 1)
-        {
-            std::ostringstream oss;
-            oss << "Conflict with more that one row: " << model;
-            logger::LOG.error(oss.str());
-
-            // Mark the connection for reset
-            stateGuard.markForReset();
-
-            throw;
-        }
-
-        if (retrieved.size() == 1 && update_on_conflict)
-        {
-            // Update
-            auto update_stmt = model.prepareUpdateStatement(t, conn);
-            auto uprep       = conn.prepare(update_stmt);
-            uprep.params.id  = retrieved[0].getIdAsString();
-            conn(uprep);
-        }
-        else
-        {
-            // Insert
-            auto insert_stmt = model.prepareInsertStatement(t, conn);
-            conn(insert_stmt);
-        }
-    }
-    catch (const std::exception& e)
-    {
-        std::ostringstream oss;
-        oss << "Error saving " << ModelType::modelName() << ": " << e.what();
-        logger::LOG.error(oss.str());
-
-        // Mark the connection for reset
-        stateGuard.markForReset();
-
-        throw;
-    }
-}
+void save(ModelType& model, std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const bool update_on_conflict);
 
 // TODO: Implement batch save for models.
 
 template < typename ModelType >
-void batchSave(const std::vector< ModelType >& models, std::shared_ptr< sqlpp::postgresql::connection > conn_ptr)
-{
-    if (models.empty())
-    {
-        return;
-    }
-
-    // Use the provided connection if available, otherwise get the default connection
-    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
-    const auto& t = ModelType::table();
-
-    // Create state guard for this connection
-    ConnectionStateGuard stateGuard(conn);
-
-    try
-    {
-        auto batch_insert_stmt = ModelType::prepareBatchInsertStatement(models, t, conn);
-        conn(batch_insert_stmt);
-    }
-    catch (const std::exception& e)
-    {
-        std::ostringstream oss;
-        oss << "Error in batch saving " << ModelType::modelName() << ": " << e.what();
-        logger::LOG.error(oss.str());
-
-        // Mark the connection for reset
-        stateGuard.markForReset();
-
-        throw;
-    }
-}
+void batchSave(const std::vector< ModelType >& models, std::shared_ptr< sqlpp::postgresql::connection > conn_ptr);
 
 namespace order
 {
@@ -1211,27 +574,7 @@ class Order
           std::optional< int64_t > executed_at,
           std::optional< int64_t > canceled_at,
           nlohmann::json vars,
-          enums::OrderSubmittedVia submitted_via)
-        : id_(boost::uuids::random_generator()())
-        , trade_id_(trade_id)
-        , session_id_(session_id)
-        , exchange_id_(exchange_id)
-        , symbol_(symbol)
-        , exchange_name_(exchange_name)
-        , order_side_(order_side)
-        , order_type_(order_type)
-        , reduce_only_(reduce_only)
-        , qty_(qty)
-        , filled_qty_(filled_qty)
-        , price_(price)
-        , status_(status)
-        , created_at_(created_at)
-        , executed_at_(executed_at)
-        , canceled_at_(canceled_at)
-        , vars_(vars)
-        , submitted_via_(submitted_via)
-    {
-    }
+          enums::OrderSubmittedVia submitted_via);
 
     // Rule of five
     Order(const Order&)                = default;
@@ -1330,245 +673,23 @@ class Order
     //     return selectors.get_position(self.exchange, self.symbol)
 
     // Calculated properties
-    double getValue() const
-    {
-        if (!price_)
-        {
-            return 0.0;
-        }
-        return std::abs(qty_) * (*price_);
-    }
+    double getValue() const;
 
-    double getRemainingQty() const
-    {
-        return helper::prepareQty(std::abs(qty_) - std::abs(filled_qty_), enums::toString(order_side_));
-    }
+    double getRemainingQty() const;
 
     // Order state transitions
-    void queueIt()
-    {
-        status_ = enums::OrderStatus::QUEUED;
-        canceled_at_.reset();
+    void queueIt();
 
-        if (helper::isDebuggable("order_submission"))
-        {
-            std::string txt = "QUEUED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
-                              enums::toString(order_side_) + ", " + std::to_string(qty_);
+    void resubmit();
 
-            if (price_)
-            {
-                txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
-            }
+    void cancel(bool silent = false, const std::string& source = "");
 
-            logger::LOG.info(txt);
-        }
+    void execute(bool silent = false);
 
-        notifySubmission();
-    }
-
-    void resubmit()
-    {
-        // Don't allow resubmission if the order is not queued
-        if (!isQueued())
-        {
-            throw std::runtime_error("Cannot resubmit an order that is not queued. Current status: " +
-                                     enums::toString(status_));
-        }
-
-        // Regenerate the order id to avoid errors on the exchange's side
-        id_     = boost::uuids::random_generator()();
-        status_ = enums::OrderStatus::ACTIVE;
-        canceled_at_.reset();
-
-        if (helper::isDebuggable("order_submission"))
-        {
-            std::string txt = "SUBMITTED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
-                              enums::toString(order_side_) + ", " + std::to_string(qty_);
-
-            if (price_)
-            {
-                txt += ", $" + std::to_string(*price_);
-            }
-
-            logger::LOG.info(txt);
-        }
-
-        notifySubmission();
-    }
-
-    void cancel(bool silent = false, const std::string& source = "")
-    {
-        if (isCanceled() || isExecuted())
-        {
-            return;
-        }
-
-        // Fix for when the cancelled stream's lag causes cancellation of queued orders
-        if (source == "stream" && isQueued())
-        {
-            return;
-        }
-
-        canceled_at_ = helper::nowToTimestamp();
-        status_      = enums::OrderStatus::CANCELED;
-
-        // TODO:
-        // if (helper::isLive())
-        // {
-        //     save();
-        // }
-
-        if (!silent)
-        {
-            std::string txt = "CANCELED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
-                              enums::toString(order_side_) + ", " + std::to_string(qty_);
-
-            if (price_)
-            {
-                txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
-            }
-
-            if (helper::isDebuggable("order_cancellation"))
-            {
-                logger::LOG.info(txt);
-            }
-
-            if (helper::isLive())
-            {
-                if (config::Config::getInstance().getValue< bool >("env_notifications_events_cancelled_orders"))
-                {
-                    // TODO:
-                    // notify(txt);
-                }
-            }
-        }
-
-        // TODO: Handle exchange balance
-        // auto e = selectors.get_exchange(exchange_);
-        // e.on_order_cancellation(*this);
-    }
-
-    void execute(bool silent = false)
-    {
-        if (isCanceled() || isExecuted())
-        {
-            return;
-        }
-
-        executed_at_ = helper::nowToTimestamp();
-        status_      = enums::OrderStatus::EXECUTED;
-
-        // TODO:
-        // if (helper::isLive())
-        // {
-        //     save();
-        // }
-
-        if (!silent)
-        {
-            std::string txt = "EXECUTED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
-                              enums::toString(order_side_) + ", " + std::to_string(qty_);
-
-            if (price_)
-            {
-                txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
-            }
-
-            if (helper::isDebuggable("order_execution"))
-            {
-                logger::LOG.info(txt);
-            }
-
-            if (helper::isLive())
-            {
-                if (config::Config::getInstance().getValue< bool >("env_notifications_events_executed_orders"))
-                {
-                    // TODO:
-                    // notify(txt);
-                }
-            }
-        }
-
-        // TODO: Log the order of the trade for metrics
-        // store.completed_trades.add_executed_order(*this);
-
-        // TODO: Handle exchange balance
-        // auto e = selectors.get_exchange(exchange_);
-        // e.on_order_execution(*this);
-
-        // TODO: Update position
-        // auto p = selectors.get_position(exchange_, symbol_);
-        // if (p) {
-        //     p._on_executed_order(*this);
-        // }
-    }
-
-    void executePartially(bool silent = false)
-    {
-        executed_at_ = helper::nowToTimestamp();
-        status_      = enums::OrderStatus::PARTIALLY_FILLED;
-
-        // TODO:
-        // if (helper::isLive())
-        // {
-        //     save();
-        // }
-
-        if (!silent)
-        {
-            std::string txt = "PARTIALLY FILLED: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
-                              enums::toString(order_side_) + ", filled qty: " + std::to_string(filled_qty_) +
-                              ", remaining qty: " + std::to_string(getRemainingQty());
-
-            if (price_)
-            {
-                txt += ", price: " + std::to_string(*price_);
-            }
-
-            if (helper::isDebuggable("order_execution"))
-            {
-                logger::LOG.info(txt);
-            }
-
-            if (helper::isLive())
-            {
-                if (config::Config::getInstance().getValue< bool >("env_notifications_events_executed_orders"))
-                {
-                    // TODO:
-                    // notify(txt);
-                }
-            }
-        }
-
-        // TODO: Log the order of the trade for metrics
-        // store.completed_trades.add_executed_order(*this);
-
-        // TODO: Update position
-        // auto p = selectors.get_position(exchange_, symbol_);
-        // if (p) {
-        //     p._on_executed_order(*this);
-        // }
-    }
+    void executePartially(bool silent = false);
 
     // Notification methods
-    void notifySubmission() const
-    {
-        if (config::Config::getInstance().getValue< bool >("env_notifications_events_submitted_orders") &&
-            (isActive() || isQueued()))
-        {
-            std::string txt = (isQueued() ? "QUEUED" : "SUBMITTED") + std::string(" order: ") + symbol_ + ", " +
-                              enums::toString(order_type_) + ", " + enums::toString(order_side_) + ", " +
-                              std::to_string(qty_);
-
-            if (price_)
-            {
-                txt += ", $" + std::to_string(*price_);
-            }
-
-            // TODO: Use notify function when available
-            std::cout << "NOTIFICATION: " << txt << std::endl;
-        }
-    }
+    void notifySubmission() const;
 
     /**
      * Creates a fake order with optional custom attributes for testing purposes.
@@ -1576,264 +697,24 @@ class Order
      * @param attributes Optional map of attributes to override defaults
      * @return Order A fake order instance
      */
-    static Order generateFakeOrder(const std::unordered_map< std::string, std::any >& attributes = {})
-    {
-        static int64_t first_timestamp = 1552309186171;
-        first_timestamp += 60000;
-
-        // Default values
-        auto exchange_name        = enums::ExchangeName::SANDBOX;
-        std::string symbol        = "BTC-USD";
-        auto order_side           = enums::OrderSide::BUY;
-        auto order_type           = enums::OrderType::LIMIT;
-        double price              = candle::randint(40, 100);
-        double qty                = candle::randint(1, 10);
-        enums::OrderStatus status = enums::OrderStatus::ACTIVE;
-        int64_t created_at        = first_timestamp;
-
-        // Prepare attributes map with defaults
-        std::unordered_map< std::string, std::any > order_attrs;
-
-        // Set the ID
-        order_attrs["id"] = boost::uuids::random_generator()();
-
-        // Set attributes with values from the provided map or use defaults
-        auto tryGet = [&attributes](const std::string& key, const auto& default_value) -> auto
-        {
-            auto it = attributes.find(key);
-            if (it != attributes.end())
-            {
-                try
-                {
-                    return std::any_cast< std::decay_t< decltype(default_value) > >(it->second);
-                }
-                catch (const std::bad_any_cast&)
-                {
-                    return default_value;
-                }
-            }
-            return default_value;
-        };
-
-        order_attrs["symbol"]        = tryGet("symbol", symbol);
-        order_attrs["exchange_name"] = tryGet("exchange_name", exchange_name);
-        order_attrs["order_side"]    = tryGet("order_side", order_side);
-        order_attrs["order_type"]    = tryGet("order_type", order_type);
-        order_attrs["qty"]           = tryGet("qty", qty);
-        order_attrs["price"]         = tryGet("price", price);
-        order_attrs["status"]        = tryGet("status", status);
-        order_attrs["created_at"]    = tryGet("created_at", created_at);
-
-        // Create the order
-        return Order(order_attrs);
-    }
+    static Order generateFakeOrder(const std::unordered_map< std::string, std::any >& attributes = {});
 
     // Database operations
     static inline auto table() { return OrdersTable{}; }
     static inline std::string modelName() { return "Order"; }
 
     template < typename ROW >
-    static Order fromRow(const ROW& row)
-    {
-        Order order;
-        order.id_ = boost::uuids::string_generator()(row.id.value());
+    static Order fromRow(const ROW& row);
 
-        if (!row.trade_id.is_null())
-            order.trade_id_ = boost::uuids::string_generator()(row.trade_id.value());
+    auto prepareSelectStatementForConflictCheck(const OrdersTable& t, sqlpp::postgresql::connection& conn) const;
 
-        order.session_id_ = boost::uuids::string_generator()(row.session_id.value());
+    auto prepareInsertStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const;
 
-        if (!row.exchange_id.is_null())
-            order.exchange_id_ = row.exchange_id.value();
-
-        order.symbol_        = row.symbol;
-        order.exchange_name_ = enums::toExchangeName(row.exchange_name);
-        order.order_side_    = enums::toOrderSide(row.order_side);
-        order.order_type_    = enums::toOrderType(row.order_type);
-        order.reduce_only_   = row.reduce_only;
-        order.qty_           = row.qty;
-        order.filled_qty_    = row.filled_qty;
-
-        if (!row.price.is_null())
-            order.price_ = row.price.value();
-
-        order.status_     = enums::toOrderStatus(row.status);
-        order.created_at_ = row.created_at;
-
-        if (!row.executed_at.is_null())
-            order.executed_at_ = row.executed_at.value();
-
-        if (!row.canceled_at.is_null())
-            order.canceled_at_ = row.canceled_at.value();
-
-        // Parse JSON from string
-        order.vars_ = nlohmann::json::parse(row.vars.value());
-
-        return order;
-    }
-
-    auto prepareSelectStatementForConflictCheck(const OrdersTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        auto query  = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
-        auto filter = createFilter()
-                          .withTradeId(trade_id_.value_or(boost::uuids::nil_uuid()))
-                          .withExchangeName(exchange_name_)
-                          .withSymbol(symbol_)
-                          .withStatus(status_)
-                          .withCreatedAt(created_at_);
-
-        filter.applyToQuery(query, t);
-
-        return query;
-    }
-
-    auto prepareInsertStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        auto stmt = sqlpp::dynamic_insert_into(conn, t).dynamic_set(t.id            = getIdAsString(),
-                                                                    t.session_id    = getSessionIdAsString(),
-                                                                    t.symbol        = symbol_,
-                                                                    t.exchange_name = enums::toString(exchange_name_),
-                                                                    t.order_side    = enums::toString(order_side_),
-                                                                    t.order_type    = enums::toString(order_type_),
-                                                                    t.reduce_only   = reduce_only_,
-                                                                    t.qty           = qty_,
-                                                                    t.filled_qty    = filled_qty_,
-                                                                    t.status        = enums::toString(status_),
-                                                                    t.created_at    = created_at_,
-                                                                    t.vars          = vars_.dump());
-
-        // Add optional fields
-        if (trade_id_ && *trade_id_ != boost::uuids::nil_uuid())
-        {
-            stmt.insert_list.add(t.trade_id = getTradeIdAsString());
-        }
-
-        if (exchange_id_ && *exchange_id_ != "")
-        {
-            stmt.insert_list.add(t.exchange_id = *exchange_id_);
-        }
-
-        if (price_ && *price_ != std::numeric_limits< double >::quiet_NaN())
-        {
-            stmt.insert_list.add(t.price = *price_);
-        }
-
-        if (executed_at_ && *executed_at_ != std::numeric_limits< int64_t >::min())
-        {
-            stmt.insert_list.add(t.executed_at = *executed_at_);
-        }
-
-        if (canceled_at_ && canceled_at_ != std::numeric_limits< int64_t >::min())
-        {
-            stmt.insert_list.add(t.canceled_at = *canceled_at_);
-        }
-
-        return stmt;
-    }
-
-    // TODO: performance.
     auto prepareBatchInsertStatement(const std::vector< Order >& models,
                                      const OrdersTable& t,
-                                     sqlpp::postgresql::connection& conn)
-    {
-        if (models.empty())
-        {
-            throw std::invalid_argument("Cannot prepare batch insert for empty models vector");
-        }
+                                     sqlpp::postgresql::connection& conn);
 
-        auto stmt = models[0].prepareInsertStatement(t, conn);
-
-        std::vector< decltype(stmt) > stmts;
-        stmts.reserve(models.size());
-        stmts.emplace_back(stmt);
-
-        for (size_t i = 1; i < models.size(); ++i)
-        {
-            stmts.emplace_back(models[i].prepareInsertStatement(t, conn));
-        }
-
-        return stmts;
-    }
-
-    auto prepareUpdateStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        auto stmt = sqlpp::dynamic_update(conn, t)
-                        .dynamic_set(t.session_id    = getSessionIdAsString(),
-                                     t.symbol        = symbol_,
-                                     t.exchange_name = enums::toString(exchange_name_),
-                                     t.order_side    = enums::toString(order_side_),
-                                     t.order_type    = enums::toString(order_type_),
-                                     t.reduce_only   = reduce_only_,
-                                     t.qty           = qty_,
-                                     t.filled_qty    = filled_qty_,
-                                     t.status        = enums::toString(status_),
-                                     t.created_at    = created_at_,
-                                     t.vars          = vars_.dump())
-                        .dynamic_where(t.id == parameter(t.id));
-
-        // Add optional fields
-        if (trade_id_)
-        {
-            if (*trade_id_ != boost::uuids::nil_uuid())
-            {
-                stmt.assignments.add(t.trade_id = getTradeIdAsString());
-            }
-            else
-            {
-                stmt.assignments.add(t.trade_id = sqlpp::null);
-            }
-        }
-
-        if (exchange_id_)
-        {
-            if (*exchange_id_ != "")
-            {
-                stmt.assignments.add(t.exchange_id = *exchange_id_);
-            }
-            else
-            {
-                stmt.assignments.add(t.exchange_id = sqlpp::null);
-            }
-        }
-
-        if (price_)
-        {
-            if (*price_ != std::numeric_limits< double >::quiet_NaN())
-            {
-                stmt.assignments.add(t.price = *price_);
-            }
-            else
-            {
-                stmt.assignments.add(t.price = sqlpp::null);
-            }
-        }
-
-        if (executed_at_)
-        {
-            if (*executed_at_ != std::numeric_limits< int64_t >::min())
-            {
-                stmt.assignments.add(t.executed_at = *executed_at_);
-            }
-            else
-            {
-                stmt.assignments.add(t.executed_at = sqlpp::null);
-            }
-        }
-
-        if (canceled_at_)
-        {
-            if (*canceled_at_ != std::numeric_limits< int64_t >::min())
-            {
-                stmt.assignments.add(t.canceled_at = *canceled_at_);
-            }
-            else
-            {
-                stmt.assignments.add(t.canceled_at = sqlpp::null);
-            }
-        }
-
-        return stmt;
-    }
+    auto prepareUpdateStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const;
 
     void save(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr = nullptr, bool update_on_conflict = false)
     {
@@ -1844,63 +725,6 @@ class Order
                                            const boost::uuids::uuid& id)
     {
         return db::findById< Order >(conn_ptr, id);
-    }
-
-    // Return a dictionary representation of the order
-    nlohmann::json toJson() const
-    {
-        nlohmann::json result;
-
-        result["id"]         = getIdAsString();
-        result["trade_id"]   = getTradeIdAsString();
-        result["session_id"] = getSessionIdAsString();
-
-        if (exchange_id_)
-        {
-            result["exchange_id"] = *exchange_id_;
-        }
-        else
-        {
-            result["exchange_id"] = nullptr;
-        }
-
-        result["symbol"]     = symbol_;
-        result["order_side"] = order_side_;
-        result["order_type"] = order_type_;
-        result["qty"]        = qty_;
-        result["filled_qty"] = filled_qty_;
-
-        if (price_)
-        {
-            result["price"] = *price_;
-        }
-        else
-        {
-            result["price"] = nullptr;
-        }
-
-        result["status"]     = status_;
-        result["created_at"] = created_at_;
-
-        if (canceled_at_)
-        {
-            result["canceled_at"] = *canceled_at_;
-        }
-        else
-        {
-            result["canceled_at"] = nullptr;
-        }
-
-        if (executed_at_)
-        {
-            result["executed_at"] = *executed_at_;
-        }
-        else
-        {
-            result["executed_at"] = nullptr;
-        }
-
-        return result;
     }
 
     // Query builder for flexible filtering
@@ -1962,61 +786,7 @@ class Order
         }
 
         template < typename Query, typename Table >
-        void applyToQuery(Query& query, const Table& t) const
-        {
-            if (id_)
-            {
-                query.where.add(t.id == boost::uuids::to_string(*id_));
-            }
-
-            if (trade_id_)
-            {
-                if (*trade_id_ != boost::uuids::nil_uuid())
-                {
-                    query.where.add(t.trade_id == boost::uuids::to_string(*trade_id_));
-                }
-                else
-                {
-                    query.where.add(t.trade_id.is_null());
-                }
-            }
-
-            if (session_id_)
-            {
-                query.where.add(t.session_id == boost::uuids::to_string(*session_id_));
-            }
-
-            if (symbol_)
-            {
-                query.where.add(t.symbol == *symbol_);
-            }
-
-            if (exchange_name_)
-            {
-                query.where.add(t.exchange_name == enums::toString(*exchange_name_));
-            }
-
-            if (order_side_)
-            {
-                query.where.add(t.order_side == enums::toString(*order_side_));
-            }
-
-            if (order_type_)
-            {
-                query.where.add(t.order_type == enums::toString(*order_type_));
-            }
-
-            if (status_)
-            {
-                query.where.add(t.status == enums::toString(*status_));
-            }
-
-            if (created_at_)
-            {
-                query.where.add(t.created_at == *created_at_);
-            }
-        }
-
+        void applyToQuery(Query& query, const Table& t) const;
 
        private:
         std::optional< boost::uuids::uuid > id_;
@@ -2030,14 +800,14 @@ class Order
         std::optional< int64_t > created_at_;
     };
 
-    static Filter createFilter() { return Filter{}; }
-
     static std::optional< std::vector< Order > > findByFilter(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
                                                               const Filter& filter)
-
     {
         return db::findByFilter< Order, Filter >(conn_ptr, filter);
     }
+
+    // Return a dictionary representation of the order
+    nlohmann::json toJson() const;
 
    private:
     boost::uuids::uuid id_;
@@ -2291,29 +1061,15 @@ class Candle
     Candle();
     explicit Candle(const std::unordered_map< std::string, std::any >& attributes);
 
-    Candle(
-
-        int64_t timestamp,
-        double open,
-        double close,
-        double high,
-        double low,
-        double volume,
-        enums::ExchangeName exchange_name,
-        std::string symbol,
-        enums::Timeframe timeframe)
-        : id_(boost::uuids::random_generator()())
-        , timestamp_(timestamp)
-        , open_(open)
-        , close_(close)
-        , high_(high)
-        , low_(low)
-        , volume_(volume)
-        , exchange_name_(exchange_name)
-        , symbol_(symbol)
-        , timeframe_(timeframe)
-    {
-    }
+    Candle(int64_t timestamp,
+           double open,
+           double close,
+           double high,
+           double low,
+           double volume,
+           enums::ExchangeName exchange_name,
+           std::string symbol,
+           enums::Timeframe timeframe);
 
     // Rule of five
     Candle(const Candle&)                = default;
@@ -2361,66 +1117,15 @@ class Candle
 
     // Convert DB row to model instance
     template < typename ROW >
-    static Candle fromRow(const ROW& row)
-    {
-        Candle candle;
-        candle.id_            = boost::uuids::string_generator()(row.id.value());
-        candle.timestamp_     = row.timestamp;
-        candle.open_          = row.open;
-        candle.close_         = row.close;
-        candle.high_          = row.high;
-        candle.low_           = row.low;
-        candle.volume_        = row.volume;
-        candle.exchange_name_ = enums::toExchangeName(row.exchange_name);
-        candle.symbol_        = row.symbol;
-        candle.timeframe_     = enums::toTimeframe(row.timeframe);
-        return candle;
-    }
+    static Candle fromRow(const ROW& row);
 
-    auto prepareSelectStatementForConflictCheck(const CandlesTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        auto query  = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
-        auto filter = createFilter()
-                          .withExchangeName(exchange_name_)
-                          .withSymbol(symbol_)
-                          .withTimeframe(timeframe_)
-                          .withTimestamp(timestamp_);
-
-        filter.applyToQuery(query, t);
-
-        return query;
-    }
+    auto prepareSelectStatementForConflictCheck(const CandlesTable& t, sqlpp::postgresql::connection& conn) const;
 
     // Prepare insert statement
-    auto prepareInsertStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        return sqlpp::dynamic_insert_into(conn, t).dynamic_set(t.id            = getIdAsString(),
-                                                               t.timestamp     = timestamp_,
-                                                               t.open          = open_,
-                                                               t.close         = close_,
-                                                               t.high          = high_,
-                                                               t.low           = low_,
-                                                               t.volume        = volume_,
-                                                               t.exchange_name = enums::toString(exchange_name_),
-                                                               t.symbol        = symbol_,
-                                                               t.timeframe     = enums::toString(timeframe_));
-    }
+    auto prepareInsertStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const;
 
     // Prepare update statement
-    auto prepareUpdateStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const
-    {
-        return sqlpp::dynamic_update(conn, t)
-            .dynamic_set(t.timestamp     = timestamp_,
-                         t.open          = open_,
-                         t.close         = close_,
-                         t.high          = high_,
-                         t.low           = low_,
-                         t.volume        = volume_,
-                         t.exchange_name = enums::toString(exchange_name_),
-                         t.symbol        = symbol_,
-                         t.timeframe     = enums::toString(timeframe_))
-            .dynamic_where(t.id == parameter(t.id));
-    }
+    auto prepareUpdateStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const;
 
     // Simplified public interface methods that use the generic functions
     void save(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr = nullptr, bool update_on_conflict = false)
@@ -2505,49 +1210,7 @@ class Candle
         }
 
         template < typename Query, typename Table >
-        void applyToQuery(Query& query, const Table& t) const
-        {
-            if (id_)
-            {
-                query.where.add(t.id == boost::uuids::to_string(*id_));
-            }
-            if (timestamp_)
-            {
-                query.where.add(t.timestamp == *timestamp_);
-            }
-            if (open_)
-            {
-                query.where.add(t.open == *open_);
-            }
-            if (close_)
-            {
-                query.where.add(t.close == *close_);
-            }
-            if (high_)
-            {
-                query.where.add(t.high == *high_);
-            }
-            if (low_)
-            {
-                query.where.add(t.low == *low_);
-            }
-            if (volume_)
-            {
-                query.where.add(t.volume == *volume_);
-            }
-            if (exchange_name_)
-            {
-                query.where.add(t.exchange_name == enums::toString(*exchange_name_));
-            }
-            if (symbol_)
-            {
-                query.where.add(t.symbol == *symbol_);
-            }
-            if (timeframe_)
-            {
-                query.where.add(t.timeframe == enums::toString(*timeframe_));
-            }
-        }
+        void applyToQuery(Query& query, const Table& t) const;
 
        private:
         std::optional< boost::uuids::uuid > id_;
@@ -2561,9 +1224,6 @@ class Candle
         std::optional< std::string > symbol_;
         std::optional< enums::Timeframe > timeframe_;
     };
-
-    // Static factory method for creating a filter
-    static Filter createFilter() { return Filter{}; }
 
     static std::optional< std::vector< Candle > > findByFilter(
         std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const Filter& filter)
@@ -2594,60 +1254,11 @@ class Candle
  * @param timeframe Timeframe string
  * @param candles Blaze matrix containing candle data
  */
-inline void saveCandles(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+extern void saveCandles(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
                         const enums::ExchangeName& exchange_name,
                         const std::string& symbol,
                         const enums::Timeframe& timeframe,
-                        const blaze::DynamicMatrix< double >& candles)
-{
-    // Make sure the number of candles is more than 0
-    if (candles.rows() == 0)
-    {
-        throw std::runtime_error("No candles to store for " + enums::toString(exchange_name) + "-" + symbol + "-" +
-                                 enums::toString(timeframe));
-    }
-
-    // Use the provided connection if available, otherwise get the default connection
-    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
-    const auto& t = Candle::table();
-
-    // Create state guard for this connection
-    ConnectionStateGuard stateGuard(conn);
-
-    auto stmt = sqlpp::insert_into(t).columns(
-        t.id, t.timestamp, t.open, t.close, t.high, t.low, t.volume, t.exchange_name, t.symbol, t.timeframe);
-
-    for (size_t i = 0; i < candles.rows(); ++i)
-    {
-        auto id = boost::uuids::to_string(boost::uuids::random_generator()());
-        stmt.values.add(t.id            = id,
-                        t.timestamp     = static_cast< int64_t >(candles(i, 0)),
-                        t.open          = candles(i, 1),
-                        t.close         = candles(i, 2),
-                        t.high          = candles(i, 3),
-                        t.low           = candles(i, 4),
-                        t.volume        = candles(i, 5),
-                        t.exchange_name = enums::toString(exchange_name),
-                        t.symbol        = symbol,
-                        t.timeframe     = enums::toString(timeframe));
-    }
-
-    try
-    {
-        conn(stmt);
-    }
-    catch (const std::exception& e)
-    {
-        std::ostringstream oss;
-        oss << "Error saving candles: " << e.what();
-        logger::LOG.error(oss.str());
-
-        // Mark the connection for reset
-        stateGuard.markForReset();
-
-        throw;
-    }
-}
+                        const blaze::DynamicMatrix< double >& candles);
 
 inline std::ostream& operator<<(std::ostream& os, const Candle& candle)
 {
