+++ src/DB.cpp
@@ -1,54 +1,74 @@
-#include "DB.hpp"
-#include <any>
-#include <chrono>
-#include <cmath>
-#include <cstdint>
-#include <optional>
-#include <stdexcept>
-#include <string>
-#include <unordered_map>
-#include <vector>
-#include "Config.hpp"
 #include "DynamicArray.hpp"
+
+#include "Candle.hpp"
+#include "Config.hpp"
+#include "DB.hpp"
 #include "Enum.hpp"
 #include "Helper.hpp"
 #include "Logger.hpp"
 
-// Basic sqlpp11 headers
-#include <sqlpp11/boolean_expression.h>
-#include <sqlpp11/postgresql/postgresql.h> // If using PostgreSQL
-#include <sqlpp11/sqlpp11.h>
+ct::db::DatabaseShutdownManager& ct::db::DatabaseShutdownManager::getInstance()
+{
+    static DatabaseShutdownManager instance;
+    return instance;
+}
 
-// For using parameters
-#include <sqlpp11/parameter.h>
-#include <sqlpp11/parameter_list.h>
+// Register a hook to be called during shutdown
+void ct::db::DatabaseShutdownManager::registerShutdownHook(ShutdownHook hook)
+{
+    std::lock_guard< std::mutex > lock(hooksMutex_);
+    shutdownHooks_.push_back(std::move(hook));
+}
 
-// For data types
-#include <sqlpp11/data_types/blob.h>
-#include <sqlpp11/data_types/floating_point.h>
-#include <sqlpp11/data_types/integral.h>
+// Register a hook to be called after shutdown is complete
+void ct::db::DatabaseShutdownManager::registerCompletionHook(ShutdownCompletionHook hook)
+{
+    std::lock_guard< std::mutex > lock(completionHooksMutex_);
+    completionHooks_.push_back(std::move(hook));
+}
 
-// For common operations
-#include <sqlpp11/aggregate_functions.h>
-#include <sqlpp11/functions.h>
-#include <sqlpp11/insert.h>
-#include <sqlpp11/postgresql/postgresql.h>
-#include <sqlpp11/select.h>
-#include <sqlpp11/transaction.h>
-#include <sqlpp11/update.h>
-#include <sqlpp11/where.h>
+// Initialize signal handlers
+void ct::db::DatabaseShutdownManager::initSignalHandlers()
+{
+    std::signal(SIGINT, handleSignal);
+    std::signal(SIGTERM, handleSignal);
 
-// For prepared statements (if using that approach)
-#include <sqlpp11/prepared_insert.h>
-#include <sqlpp11/prepared_select.h>
-#include <sqlpp11/prepared_update.h>
+    // Ignore SIGPIPE to prevent crashes on broken connections
+    std::signal(SIGPIPE, SIG_IGN);
+}
 
-// For Boost UUID
-#include <boost/uuid/random_generator.hpp>
-#include <boost/uuid/string_generator.hpp>
-#include <boost/uuid/uuid.hpp>
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
+// Check if shutdown is in progress
+bool ct::db::DatabaseShutdownManager::isShuttingDown() const
+{
+    return shuttingDown_.load(std::memory_order_acquire);
+}
+
+// Wait for shutdown to complete
+void ct::db::DatabaseShutdownManager::waitForShutdown()
+{
+    if (shutdownFuture_.valid())
+    {
+        shutdownFuture_.wait();
+    }
+}
+
+void ct::db::DatabaseShutdownManager::shutdown()
+{
+    // Only start shutdown once
+    bool expected = false;
+    if (!shuttingDown_.compare_exchange_strong(expected, true, std::memory_order_acq_rel))
+    {
+        return;
+    }
+
+    // Launch shutdown process asynchronously
+    shutdownFuture_ = std::async(std::launch::async, [this] { performShutdown(); });
+}
+
+void ct::db::DatabaseShutdownManager::handleSignal(int signal)
+{
+    getInstance().shutdown();
+}
 
 void ct::db::DatabaseShutdownManager::performShutdown()
 {
@@ -90,6 +110,677 @@ void ct::db::DatabaseShutdownManager::performShutdown()
                 logger::LOG.error(oss.str());
             }
         }
+    }
+}
+
+// Get singleton instance
+ct::db::ConnectionPool& ct::db::ConnectionPool::getInstance()
+{
+    static ConnectionPool instance;
+    return instance;
+}
+
+// Initialize the connection pool
+void ct::db::ConnectionPool::init(const std::string& host,
+                                  const std::string& dbname,
+                                  const std::string& username,
+                                  const std::string& password,
+                                  unsigned int port,
+                                  size_t poolSize) // Default pool size
+{
+    std::lock_guard< std::mutex > lock(mutex_);
+
+    // Store connection parameters for creating new connections
+    host_     = host;
+    dbname_   = dbname;
+    username_ = username;
+    password_ = password;
+    port_     = port;
+
+    // Initialize the pool with connections
+    for (size_t i = 0; i < poolSize; ++i)
+    {
+        createNewConnection();
+    }
+}
+
+// Get a connection from the pool (or create one if pool is empty)
+std::shared_ptr< sqlpp::postgresql::connection > ct::db::ConnectionPool::getConnection()
+{
+    std::unique_lock< std::mutex > lock(mutex_);
+
+    if (!initialized_)
+    {
+        throw std::runtime_error("Connection pool not initialized");
+    }
+
+    // Check if we're shutting down
+    if (DatabaseShutdownManager::getInstance().isShuttingDown())
+    {
+        throw std::runtime_error("Database is shutting down");
+    }
+
+    // Wait for a connection to become available
+    while (availableConnections_.empty())
+    {
+        // If we've reached max connections, wait for one to be returned
+        if (activeConnections_ >= maxConnections_)
+        {
+            // TODO: Read timeout from config.
+            if (!connectionAvailable_.wait_for(
+                    lock, std::chrono::seconds(5), [this] { return !availableConnections_.empty(); }))
+            {
+                throw std::runtime_error("Timeout waiting for connection");
+            }
+        }
+        else
+        {
+            // Create a new connection if below max limit
+            createNewConnection();
+        }
+    }
+
+    // Get a connection from the pool
+    auto conn = std::move(availableConnections_.front());
+    availableConnections_.pop();
+
+    // Save the raw pointer before transferring ownership
+    sqlpp::postgresql::connection* rawConn = conn.get();
+
+    // IMPORTANT: Add to managed connections before creating the shared_ptr
+    managedConnections_.push_back(std::move(conn));
+    activeConnections_++;
+
+    // Create a wrapper that returns the connection to the pool when it's destroyed
+    return std::shared_ptr< sqlpp::postgresql::connection >(
+        rawConn, [this](sqlpp::postgresql::connection* conn) { this->returnConnection(conn); });
+}
+
+// Get a connection from the pool with health check
+std::shared_ptr< sqlpp::postgresql::connection > ct::db::ConnectionPool::getConnectionWithHealthCheck()
+{
+    auto conn = getConnection();
+
+    // Test if the connection is still valid
+    try
+    {
+        conn->execute("SELECT 1");
+    }
+    catch (const std::exception& e)
+    {
+        // Connection is dead, get a new one
+        std::cerr << "Detected dead connection, getting a new one: " << e.what() << std::endl;
+
+        // This will release the dead connection and get a new one
+        // The connection pool's returnConnection will handle creating a replacement
+        conn = getConnection();
+    }
+
+    return conn;
+}
+
+// Set maximum number of connections
+void ct::db::ConnectionPool::setMaxConnections(size_t maxConnections)
+{
+    std::lock_guard< std::mutex > lock(mutex_);
+    maxConnections_ = maxConnections;
+}
+
+void ct::db::ConnectionPool::waitForConnectionsToClose()
+{
+    std::unique_lock< std::mutex > lock(mutex_);
+
+    // Wait until all connections are returned
+    while (activeConnections_ > 0)
+    {
+        connectionReturned_.wait(lock);
+    }
+}
+
+// Private constructor for singleton
+ct::db::ConnectionPool::ConnectionPool() : activeConnections_(0), maxConnections_(20), initialized_(false) {}
+
+// Return a connection to the pool
+void ct::db::ConnectionPool::returnConnection(sqlpp::postgresql::connection* conn)
+{
+    std::lock_guard< std::mutex > lock(mutex_);
+
+    // Check if connection is still valid before trying to clean it up
+    bool connectionValid = true;
+    try
+    {
+        // A simple ping to test if connection is alive
+        conn->execute("SELECT 1");
+    }
+    catch (const std::exception& e)
+    {
+        // Connection is dead, mark it as invalid
+        connectionValid = false;
+        std::cerr << "Detected dead connection during return: " << e.what() << std::endl;
+    }
+
+    if (connectionValid)
+    {
+        // Only try to clean up if connection is still valid
+        try
+        {
+            conn->execute("DEALLOCATE ALL");
+        }
+        catch (const std::exception& e)
+        {
+            // Log the error but continue
+            std::cerr << "Error cleaning connection during return: " << e.what() << std::endl;
+            connectionValid = false;
+        }
+    }
+
+    // Find the connection in our managed connections
+    auto it = std::find_if(managedConnections_.begin(),
+                           managedConnections_.end(),
+                           [conn](const auto& managed) { return managed.get() == conn; });
+
+    if (it != managedConnections_.end())
+    {
+        if (!DatabaseShutdownManager::getInstance().isShuttingDown())
+        {
+            if (connectionValid)
+            {
+                // Return to the pool only if valid
+                availableConnections_.push(std::move(*it));
+            }
+            else
+            {
+                // Connection is invalid, create a new one to replace it
+                createNewConnection();
+            }
+        }
+
+        managedConnections_.erase(it);
+        activeConnections_--;
+
+        // Notify both waiting threads and shutdown manager
+        connectionAvailable_.notify_one();
+        connectionReturned_.notify_all();
+    }
+}
+
+// Create a new database connection
+void ct::db::ConnectionPool::createNewConnection()
+{
+    auto config      = std::make_shared< sqlpp::postgresql::connection_config >();
+    config->debug    = false; // TODO: accept as arg
+    config->host     = host_;
+    config->dbname   = dbname_;
+    config->user     = username_;
+    config->password = password_;
+    config->port     = port_;
+
+    auto conn = std::make_unique< sqlpp::postgresql::connection >(config);
+    availableConnections_.push(std::move(conn));
+    initialized_ = true;
+}
+
+// Get singleton instance
+ct::db::Database& ct::db::Database::getInstance()
+{
+    static Database instance;
+    return instance;
+}
+
+// Initialize the database with connection parameters
+void ct::db::Database::init(const std::string& host,
+                            const std::string& dbname,
+                            const std::string& username,
+                            const std::string& password,
+                            unsigned int port)
+{
+    std::lock_guard< std::mutex > lock(mutex_);
+
+    // Initialize connection pool
+    ConnectionPool::getInstance().init(host, dbname, username, password, port);
+
+    // Initialize shutdown manager
+    auto& shutdownManager = DatabaseShutdownManager::getInstance();
+    shutdownManager.initSignalHandlers();
+
+    // Register shutdown hooks
+    shutdownManager.registerShutdownHook([] { std::cout << "Database shutdown initiated..." << std::endl; });
+
+    shutdownManager.registerCompletionHook([] { std::cout << "Database shutdown completed." << std::endl; });
+}
+
+// Get a connection from the pool
+// sqlpp::postgresql::connection& ct::db::Database::getConnection()
+// {
+//     // Get connection from pool and cache it in thread_local storage
+//     thread_local std::shared_ptr< sqlpp::postgresql::connection > conn =
+//     ConnectionPool::getInstance().getConnection();
+
+// return *conn;
+// }
+
+std::shared_ptr< sqlpp::postgresql::connection > ct::db::Database::getConnection()
+{
+    return ConnectionPool::getInstance().getConnection();
+}
+
+void ct::db::Database::shutdown()
+{
+    DatabaseShutdownManager::getInstance().shutdown();
+    DatabaseShutdownManager::getInstance().waitForShutdown();
+}
+
+// Start a new transaction
+std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > ct::db::TransactionManager::startTransaction()
+{
+    auto db = db::Database::getInstance().getConnection();
+    return std::make_shared< sqlpp::transaction_t< sqlpp::postgresql::connection > >(start_transaction(*db));
+}
+
+// Commit the transaction
+bool ct::db::TransactionManager::commitTransaction(
+    std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx)
+{
+    if (!tx)
+    {
+        logger::LOG.error("Cannot commit null transaction");
+        return false;
+    }
+
+    try
+    {
+        tx->commit();
+        return true;
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error committing transaction: " << e.what();
+        logger::LOG.error(oss.str());
+        return false;
+    }
+}
+
+// Roll back the transaction
+bool ct::db::TransactionManager::rollbackTransaction(
+    std::shared_ptr< sqlpp::transaction_t< sqlpp::postgresql::connection > > tx)
+{
+    if (!tx)
+    {
+        std::ostringstream oss;
+        oss << "Cannot rollback null transaction";
+        logger::LOG.error(oss.str());
+        return false;
+    }
+
+    try
+    {
+        tx->rollback();
+        return true;
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error rolling back transaction: " << e.what();
+        logger::LOG.error(oss.str());
+        return false;
+    }
+}
+
+ct::db::TransactionGuard::TransactionGuard() : committed_(false)
+{
+    // Check for shutdown before starting transaction
+    if (DatabaseShutdownManager::getInstance().isShuttingDown())
+    {
+        throw std::runtime_error("Cannot start new transaction during shutdown");
+    }
+
+    conn_ = ConnectionPool::getInstance().getConnection();
+
+    if (conn_)
+    {
+        tx_ = std::make_shared< sqlpp::transaction_t< sqlpp::postgresql::connection > >(start_transaction(*conn_));
+    }
+}
+
+ct::db::TransactionGuard::~TransactionGuard()
+{
+    if (tx_ && !committed_)
+    {
+        try
+        {
+            tx_->rollback();
+        }
+        catch (const std::exception& e)
+        {
+            std::ostringstream oss;
+            oss << "Error during auto-rollback: " << e.what();
+            logger::LOG.error(oss.str());
+        }
+    }
+}
+
+bool ct::db::TransactionGuard::commit()
+{
+    if (!tx_)
+        return false;
+
+    try
+    {
+        tx_->commit();
+        committed_ = true;
+        return true;
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error during commit: " << e.what();
+        logger::LOG.error(oss.str());
+        return false;
+    }
+}
+
+bool ct::db::TransactionGuard::rollback()
+{
+    if (!tx_)
+        return false;
+
+    try
+    {
+        tx_->rollback();
+        return true;
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error during rollback: " << e.what();
+        logger::LOG.error(oss.str());
+        return false;
+    }
+}
+
+// Get the connection associated with this transaction
+std::shared_ptr< sqlpp::postgresql::connection > ct::db::TransactionGuard::getConnection()
+{
+    return conn_;
+}
+
+template < typename Func >
+auto ct::db::executeWithRetry(Func&& operation, int maxRetries) -> decltype(operation())
+{
+    int retries = 0;
+    while (true)
+    {
+        try
+        {
+            return operation();
+        }
+        catch (const std::exception& e)
+        {
+            if (++retries > maxRetries)
+            {
+                throw; // Re-throw after max retries
+            }
+
+            std::ostringstream oss;
+            oss << "Operation failed, retrying (" << retries << "/" << maxRetries << "): " << e.what();
+            logger::LOG.error(oss.str());
+
+            // Exponential backoff
+            std::this_thread::sleep_for(std::chrono::milliseconds(100 * retries));
+        }
+    }
+}
+
+ct::db::ConnectionStateGuard::ConnectionStateGuard(sqlpp::postgresql::connection& conn)
+    : conn_(conn), needs_reset_(false)
+{
+}
+
+ct::db::ConnectionStateGuard::~ConnectionStateGuard()
+{
+    if (needs_reset_)
+    {
+        try
+        {
+            // Deallocate all prepared statements
+            conn_.execute("DEALLOCATE ALL");
+
+            // Execute a harmless query to reset the connection state
+            conn_.execute("SELECT 1");
+        }
+        catch (const std::exception& e)
+        {
+            // Nothing we can do in the destructor
+        }
+    }
+}
+
+void ct::db::ConnectionStateGuard::markForReset()
+{
+    needs_reset_ = true;
+}
+
+// Generic findById implementation
+template < typename ModelType >
+std::optional< ModelType > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                            const boost::uuids::uuid& id)
+{
+    // Use the provided connection if available, otherwise get the default connection
+    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
+    const auto& t = ModelType::table();
+
+    // Create state guard for this connection
+    ConnectionStateGuard stateGuard(conn);
+
+    try
+    {
+        // Convert UUID to string
+        std::string uuid_str = boost::uuids::to_string(id);
+
+        // Prepare statement
+        auto prep      = conn.prepare(select(all_of(t)).from(t).where(t.id == parameter(t.id)));
+        prep.params.id = uuid_str;
+
+        // Execute
+        auto result = conn(prep);
+
+        if (result.empty())
+        {
+            return std::nullopt;
+        }
+
+        // Create and populate a new model instance
+        const auto& row = *result.begin();
+        return ModelType::fromRow(row);
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error finding " << ModelType::modelName() << " by ID: " << e.what();
+        logger::LOG.error(oss.str());
+
+        // Mark the connection for reset
+        stateGuard.markForReset();
+
+        throw;
+    }
+}
+
+// Generic findByFilter implementation
+template < typename ModelType, typename FilterType >
+std::optional< std::vector< ModelType > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const FilterType& filter)
+{
+    // Use the provided connection if available, otherwise get the default connection
+    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
+    const auto& t = ModelType::table();
+
+    // Create state guard for this connection
+    ConnectionStateGuard stateGuard(conn);
+
+    try
+    {
+        // Build dynamic query
+        auto query = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
+
+        // Apply filter conditions
+        filter.applyToQuery(query, t);
+
+        // Execute query
+        auto rows = conn(query);
+
+        // Process results
+        std::vector< ModelType > results;
+        for (const auto& row : rows)
+        {
+            try
+            {
+                results.push_back(std::move(ModelType::fromRow(row)));
+            }
+            catch (const std::exception& e)
+            {
+                std::ostringstream oss;
+                oss << "Error processing row: " << e.what();
+                logger::LOG.error(oss.str());
+
+                // Mark the connection for reset
+                stateGuard.markForReset();
+
+                throw;
+            }
+        }
+
+        return results;
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error in findByFilter for " << ModelType::modelName() << ": " << e.what();
+        logger::LOG.error(oss.str());
+
+        // Mark the connection for reset
+        stateGuard.markForReset();
+
+        throw;
+    }
+}
+
+// Generic save implementation
+template < typename ModelType >
+void ct::db::save(ModelType& model,
+                  std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                  const bool update_on_conflict)
+{
+    // Use the provided connection if available, otherwise get the default connection
+    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
+    const auto& t = ModelType::table();
+
+    // Create state guard for this connection
+    ConnectionStateGuard stateGuard(conn);
+
+    try
+    {
+        auto select_stmt = model.prepareSelectStatementForConflictCheck(t, conn);
+        auto sprep       = conn.prepare(select_stmt);
+        auto rows        = conn(sprep);
+
+        std::vector< ModelType > retrieved;
+
+        if (!rows.empty())
+        {
+            for (const auto& row : rows)
+            {
+                try
+                {
+                    retrieved.push_back(std::move(ModelType::fromRow(row)));
+                }
+                catch (const std::exception& e)
+                {
+                    std::ostringstream oss;
+                    oss << "Error processing row: " << e.what();
+                    logger::LOG.error(oss.str());
+
+                    // Mark the connection for reset
+                    stateGuard.markForReset();
+
+                    throw;
+                }
+            }
+        }
+
+        // TODO: Support batch update?
+        if (retrieved.size() > 1)
+        {
+            std::ostringstream oss;
+            oss << "Conflict with more that one row: " << model;
+            logger::LOG.error(oss.str());
+
+            // Mark the connection for reset
+            stateGuard.markForReset();
+
+            throw;
+        }
+
+        if (retrieved.size() == 1 && update_on_conflict)
+        {
+            // Update
+            auto update_stmt = model.prepareUpdateStatement(t, conn);
+            auto uprep       = conn.prepare(update_stmt);
+            uprep.params.id  = retrieved[0].getIdAsString();
+            conn(uprep);
+        }
+        else
+        {
+            // Insert
+            auto insert_stmt = model.prepareInsertStatement(t, conn);
+            conn(insert_stmt);
+        }
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error saving " << ModelType::modelName() << ": " << e.what();
+        logger::LOG.error(oss.str());
+
+        // Mark the connection for reset
+        stateGuard.markForReset();
+
+        throw;
+    }
+}
+
+template < typename ModelType >
+void ct::db::batchSave(const std::vector< ModelType >& models,
+                       std::shared_ptr< sqlpp::postgresql::connection > conn_ptr)
+{
+    if (models.empty())
+    {
+        return;
+    }
+
+    // Use the provided connection if available, otherwise get the default connection
+    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
+    const auto& t = ModelType::table();
+
+    // Create state guard for this connection
+    ConnectionStateGuard stateGuard(conn);
+
+    try
+    {
+        auto batch_insert_stmt = ModelType::prepareBatchInsertStatement(models, t, conn);
+        conn(batch_insert_stmt);
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error in batch saving " << ModelType::modelName() << ": " << e.what();
+        logger::LOG.error(oss.str());
+
+        // Mark the connection for reset
+        stateGuard.markForReset();
+
+        throw;
     }
 }
 
@@ -242,6 +933,660 @@ ct::db::Order::Order(const std::unordered_map< std::string, std::any >& attribut
     // Notify if needed (already handled in the default constructor this one calls)
 }
 
+ct::db::Order::Order(std::optional< boost::uuids::uuid > trade_id,
+                     boost::uuids::uuid session_id,
+                     std::optional< std::string > exchange_id,
+                     std::string symbol,
+                     enums::ExchangeName exchange_name,
+                     enums::OrderSide order_side,
+                     enums::OrderType order_type,
+                     bool reduce_only,
+                     double qty,
+                     double filled_qty,
+                     std::optional< double > price,
+                     enums::OrderStatus status,
+                     int64_t created_at,
+                     std::optional< int64_t > executed_at,
+                     std::optional< int64_t > canceled_at,
+                     nlohmann::json vars,
+                     enums::OrderSubmittedVia submitted_via)
+    : id_(boost::uuids::random_generator()())
+    , trade_id_(trade_id)
+    , session_id_(session_id)
+    , exchange_id_(exchange_id)
+    , symbol_(symbol)
+    , exchange_name_(exchange_name)
+    , order_side_(order_side)
+    , order_type_(order_type)
+    , reduce_only_(reduce_only)
+    , qty_(qty)
+    , filled_qty_(filled_qty)
+    , price_(price)
+    , status_(status)
+    , created_at_(created_at)
+    , executed_at_(executed_at)
+    , canceled_at_(canceled_at)
+    , vars_(vars)
+    , submitted_via_(submitted_via)
+{
+}
+
+// Calculated properties
+double ct::db::Order::getValue() const
+{
+    if (!price_)
+    {
+        return 0.0;
+    }
+    return std::abs(qty_) * (*price_);
+}
+
+double ct::db::Order::getRemainingQty() const
+{
+    return helper::prepareQty(std::abs(qty_) - std::abs(filled_qty_), enums::toString(order_side_));
+}
+
+// Order state transitions
+void ct::db::Order::queueIt()
+{
+    status_ = enums::OrderStatus::QUEUED;
+    canceled_at_.reset();
+
+    if (helper::isDebuggable("order_submission"))
+    {
+        std::string txt = "QUEUED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
+                          enums::toString(order_side_) + ", " + std::to_string(qty_);
+
+        if (price_)
+        {
+            txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
+        }
+
+        logger::LOG.info(txt);
+    }
+
+    notifySubmission();
+}
+
+void ct::db::Order::resubmit()
+{
+    // Don't allow resubmission if the order is not queued
+    if (!isQueued())
+    {
+        throw std::runtime_error("Cannot resubmit an order that is not queued. Current status: " +
+                                 enums::toString(status_));
+    }
+
+    // Regenerate the order id to avoid errors on the exchange's side
+    id_     = boost::uuids::random_generator()();
+    status_ = enums::OrderStatus::ACTIVE;
+    canceled_at_.reset();
+
+    if (helper::isDebuggable("order_submission"))
+    {
+        std::string txt = "SUBMITTED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
+                          enums::toString(order_side_) + ", " + std::to_string(qty_);
+
+        if (price_)
+        {
+            txt += ", $" + std::to_string(*price_);
+        }
+
+        logger::LOG.info(txt);
+    }
+
+    notifySubmission();
+}
+
+void ct::db::Order::cancel(bool silent, const std::string& source)
+{
+    if (isCanceled() || isExecuted())
+    {
+        return;
+    }
+
+    // Fix for when the cancelled stream's lag causes cancellation of queued orders
+    if (source == "stream" && isQueued())
+    {
+        return;
+    }
+
+    canceled_at_ = helper::nowToTimestamp();
+    status_      = enums::OrderStatus::CANCELED;
+
+    // TODO:
+    // if (helper::isLive())
+    // {
+    //     save();
+    // }
+
+    if (!silent)
+    {
+        std::string txt = "CANCELED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
+                          enums::toString(order_side_) + ", " + std::to_string(qty_);
+
+        if (price_)
+        {
+            txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
+        }
+
+        if (helper::isDebuggable("order_cancellation"))
+        {
+            logger::LOG.info(txt);
+        }
+
+        if (helper::isLive())
+        {
+            if (config::Config::getInstance().getValue< bool >("env_notifications_events_cancelled_orders"))
+            {
+                // TODO:
+                // notify(txt);
+            }
+        }
+    }
+
+    // TODO: Handle exchange balance
+    // auto e = selectors.get_exchange(exchange_);
+    // e.on_order_cancellation(*this);
+}
+
+void ct::db::Order::execute(bool silent)
+{
+    if (isCanceled() || isExecuted())
+    {
+        return;
+    }
+
+    executed_at_ = helper::nowToTimestamp();
+    status_      = enums::OrderStatus::EXECUTED;
+
+    // TODO:
+    // if (helper::isLive())
+    // {
+    //     save();
+    // }
+
+    if (!silent)
+    {
+        std::string txt = "EXECUTED order: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
+                          enums::toString(order_side_) + ", " + std::to_string(qty_);
+
+        if (price_)
+        {
+            txt += ", $" + std::to_string(std::round(*price_ * 100) / 100);
+        }
+
+        if (helper::isDebuggable("order_execution"))
+        {
+            logger::LOG.info(txt);
+        }
+
+        if (helper::isLive())
+        {
+            if (config::Config::getInstance().getValue< bool >("env_notifications_events_executed_orders"))
+            {
+                // TODO:
+                // notify(txt);
+            }
+        }
+    }
+
+    // TODO: Log the order of the trade for metrics
+    // store.completed_trades.add_executed_order(*this);
+
+    // TODO: Handle exchange balance
+    // auto e = selectors.get_exchange(exchange_);
+    // e.on_order_execution(*this);
+
+    // TODO: Update position
+    // auto p = selectors.get_position(exchange_, symbol_);
+    // if (p) {
+    //     p._on_executed_order(*this);
+    // }
+}
+
+void ct::db::Order::executePartially(bool silent)
+{
+    executed_at_ = helper::nowToTimestamp();
+    status_      = enums::OrderStatus::PARTIALLY_FILLED;
+
+    // TODO:
+    // if (helper::isLive())
+    // {
+    //     save();
+    // }
+
+    if (!silent)
+    {
+        std::string txt = "PARTIALLY FILLED: " + symbol_ + ", " + enums::toString(order_type_) + ", " +
+                          enums::toString(order_side_) + ", filled qty: " + std::to_string(filled_qty_) +
+                          ", remaining qty: " + std::to_string(getRemainingQty());
+
+        if (price_)
+        {
+            txt += ", price: " + std::to_string(*price_);
+        }
+
+        if (helper::isDebuggable("order_execution"))
+        {
+            logger::LOG.info(txt);
+        }
+
+        if (helper::isLive())
+        {
+            if (config::Config::getInstance().getValue< bool >("env_notifications_events_executed_orders"))
+            {
+                // TODO:
+                // notify(txt);
+            }
+        }
+    }
+
+    // TODO: Log the order of the trade for metrics
+    // store.completed_trades.add_executed_order(*this);
+
+    // TODO: Update position
+    // auto p = selectors.get_position(exchange_, symbol_);
+    // if (p) {
+    //     p._on_executed_order(*this);
+    // }
+}
+
+// Notification methods
+void ct::db::Order::notifySubmission() const
+{
+    if (config::Config::getInstance().getValue< bool >("env_notifications_events_submitted_orders") &&
+        (isActive() || isQueued()))
+    {
+        std::string txt = (isQueued() ? "QUEUED" : "SUBMITTED") + std::string(" order: ") + symbol_ + ", " +
+                          enums::toString(order_type_) + ", " + enums::toString(order_side_) + ", " +
+                          std::to_string(qty_);
+
+        if (price_)
+        {
+            txt += ", $" + std::to_string(*price_);
+        }
+
+        // TODO: Use notify function when available
+        std::cout << "NOTIFICATION: " << txt << std::endl;
+    }
+}
+
+/**
+ * Creates a fake order with optional custom attributes for testing purposes.
+ *
+ * @param attributes Optional map of attributes to override defaults
+ * @return Order A fake order instance
+ */
+ct::db::Order ct::db::Order::generateFakeOrder(const std::unordered_map< std::string, std::any >& attributes)
+{
+    static int64_t first_timestamp = 1552309186171;
+    first_timestamp += 60000;
+
+    // Default values
+    auto exchange_name        = enums::ExchangeName::SANDBOX;
+    std::string symbol        = "BTC-USD";
+    auto order_side           = enums::OrderSide::BUY;
+    auto order_type           = enums::OrderType::LIMIT;
+    double price              = ct::candle::randint(40, 100);
+    double qty                = ct::candle::randint(1, 10);
+    enums::OrderStatus status = enums::OrderStatus::ACTIVE;
+    int64_t created_at        = first_timestamp;
+
+    // Prepare attributes map with defaults
+    std::unordered_map< std::string, std::any > order_attrs;
+
+    // Set the ID
+    order_attrs["id"] = boost::uuids::random_generator()();
+
+    // Set attributes with values from the provided map or use defaults
+    auto tryGet = [&attributes](const std::string& key, const auto& default_value) -> auto
+    {
+        auto it = attributes.find(key);
+        if (it != attributes.end())
+        {
+            try
+            {
+                return std::any_cast< std::decay_t< decltype(default_value) > >(it->second);
+            }
+            catch (const std::bad_any_cast&)
+            {
+                return default_value;
+            }
+        }
+        return default_value;
+    };
+
+    order_attrs["symbol"]        = tryGet("symbol", symbol);
+    order_attrs["exchange_name"] = tryGet("exchange_name", exchange_name);
+    order_attrs["order_side"]    = tryGet("order_side", order_side);
+    order_attrs["order_type"]    = tryGet("order_type", order_type);
+    order_attrs["qty"]           = tryGet("qty", qty);
+    order_attrs["price"]         = tryGet("price", price);
+    order_attrs["status"]        = tryGet("status", status);
+    order_attrs["created_at"]    = tryGet("created_at", created_at);
+
+    // Create the order
+    return Order(order_attrs);
+}
+
+
+template < typename ROW >
+ct::db::Order ct::db::Order::fromRow(const ROW& row)
+{
+    Order order;
+    order.id_ = boost::uuids::string_generator()(row.id.value());
+
+    if (!row.trade_id.is_null())
+        order.trade_id_ = boost::uuids::string_generator()(row.trade_id.value());
+
+    order.session_id_ = boost::uuids::string_generator()(row.session_id.value());
+
+    if (!row.exchange_id.is_null())
+        order.exchange_id_ = row.exchange_id.value();
+
+    order.symbol_        = row.symbol;
+    order.exchange_name_ = enums::toExchangeName(row.exchange_name);
+    order.order_side_    = enums::toOrderSide(row.order_side);
+    order.order_type_    = enums::toOrderType(row.order_type);
+    order.reduce_only_   = row.reduce_only;
+    order.qty_           = row.qty;
+    order.filled_qty_    = row.filled_qty;
+
+    if (!row.price.is_null())
+        order.price_ = row.price.value();
+
+    order.status_     = enums::toOrderStatus(row.status);
+    order.created_at_ = row.created_at;
+
+    if (!row.executed_at.is_null())
+        order.executed_at_ = row.executed_at.value();
+
+    if (!row.canceled_at.is_null())
+        order.canceled_at_ = row.canceled_at.value();
+
+    // Parse JSON from string
+    order.vars_ = nlohmann::json::parse(row.vars.value());
+
+    return order;
+}
+
+auto ct::db::Order::prepareSelectStatementForConflictCheck(const OrdersTable& t,
+                                                           sqlpp::postgresql::connection& conn) const
+{
+    auto query  = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
+    auto filter = Filter()
+                      .withTradeId(trade_id_.value_or(boost::uuids::nil_uuid()))
+                      .withExchangeName(exchange_name_)
+                      .withSymbol(symbol_)
+                      .withStatus(status_)
+                      .withCreatedAt(created_at_);
+
+    filter.applyToQuery(query, t);
+
+    return query;
+}
+
+auto ct::db::Order::prepareInsertStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const
+{
+    auto stmt = sqlpp::dynamic_insert_into(conn, t).dynamic_set(t.id            = getIdAsString(),
+                                                                t.session_id    = getSessionIdAsString(),
+                                                                t.symbol        = symbol_,
+                                                                t.exchange_name = enums::toString(exchange_name_),
+                                                                t.order_side    = enums::toString(order_side_),
+                                                                t.order_type    = enums::toString(order_type_),
+                                                                t.reduce_only   = reduce_only_,
+                                                                t.qty           = qty_,
+                                                                t.filled_qty    = filled_qty_,
+                                                                t.status        = enums::toString(status_),
+                                                                t.created_at    = created_at_,
+                                                                t.vars          = vars_.dump());
+
+    // Add optional fields
+    if (trade_id_ && *trade_id_ != boost::uuids::nil_uuid())
+    {
+        stmt.insert_list.add(t.trade_id = getTradeIdAsString());
+    }
+
+    if (exchange_id_ && *exchange_id_ != "")
+    {
+        stmt.insert_list.add(t.exchange_id = *exchange_id_);
+    }
+
+    if (price_ && *price_ != std::numeric_limits< double >::quiet_NaN())
+    {
+        stmt.insert_list.add(t.price = *price_);
+    }
+
+    if (executed_at_ && *executed_at_ != std::numeric_limits< int64_t >::min())
+    {
+        stmt.insert_list.add(t.executed_at = *executed_at_);
+    }
+
+    if (canceled_at_ && canceled_at_ != std::numeric_limits< int64_t >::min())
+    {
+        stmt.insert_list.add(t.canceled_at = *canceled_at_);
+    }
+
+    return stmt;
+}
+
+auto ct::db::Order::prepareBatchInsertStatement(const std::vector< Order >& models,
+                                                const OrdersTable& t,
+                                                sqlpp::postgresql::connection& conn)
+{
+    if (models.empty())
+    {
+        throw std::invalid_argument("Cannot prepare batch insert for empty models vector");
+    }
+
+    auto stmt = models[0].prepareInsertStatement(t, conn);
+
+    std::vector< decltype(stmt) > stmts;
+    stmts.reserve(models.size());
+    stmts.emplace_back(stmt);
+
+    for (size_t i = 1; i < models.size(); ++i)
+    {
+        stmts.emplace_back(models[i].prepareInsertStatement(t, conn));
+    }
+
+    return stmts;
+}
+
+auto ct::db::Order::prepareUpdateStatement(const OrdersTable& t, sqlpp::postgresql::connection& conn) const
+{
+    auto stmt = sqlpp::dynamic_update(conn, t)
+                    .dynamic_set(t.session_id    = getSessionIdAsString(),
+                                 t.symbol        = symbol_,
+                                 t.exchange_name = enums::toString(exchange_name_),
+                                 t.order_side    = enums::toString(order_side_),
+                                 t.order_type    = enums::toString(order_type_),
+                                 t.reduce_only   = reduce_only_,
+                                 t.qty           = qty_,
+                                 t.filled_qty    = filled_qty_,
+                                 t.status        = enums::toString(status_),
+                                 t.created_at    = created_at_,
+                                 t.vars          = vars_.dump())
+                    .dynamic_where(t.id == parameter(t.id));
+
+    // Add optional fields
+    if (trade_id_)
+    {
+        if (*trade_id_ != boost::uuids::nil_uuid())
+        {
+            stmt.assignments.add(t.trade_id = getTradeIdAsString());
+        }
+        else
+        {
+            stmt.assignments.add(t.trade_id = sqlpp::null);
+        }
+    }
+
+    if (exchange_id_)
+    {
+        if (*exchange_id_ != "")
+        {
+            stmt.assignments.add(t.exchange_id = *exchange_id_);
+        }
+        else
+        {
+            stmt.assignments.add(t.exchange_id = sqlpp::null);
+        }
+    }
+
+    if (price_)
+    {
+        if (*price_ != std::numeric_limits< double >::quiet_NaN())
+        {
+            stmt.assignments.add(t.price = *price_);
+        }
+        else
+        {
+            stmt.assignments.add(t.price = sqlpp::null);
+        }
+    }
+
+    if (executed_at_)
+    {
+        if (*executed_at_ != std::numeric_limits< int64_t >::min())
+        {
+            stmt.assignments.add(t.executed_at = *executed_at_);
+        }
+        else
+        {
+            stmt.assignments.add(t.executed_at = sqlpp::null);
+        }
+    }
+
+    if (canceled_at_)
+    {
+        if (*canceled_at_ != std::numeric_limits< int64_t >::min())
+        {
+            stmt.assignments.add(t.canceled_at = *canceled_at_);
+        }
+        else
+        {
+            stmt.assignments.add(t.canceled_at = sqlpp::null);
+        }
+    }
+
+    return stmt;
+}
+
+template < typename Query, typename Table >
+void ct::db::Order::Filter::applyToQuery(Query& query, const Table& t) const
+{
+    if (id_)
+    {
+        query.where.add(t.id == boost::uuids::to_string(*id_));
+    }
+
+    if (trade_id_)
+    {
+        if (*trade_id_ != boost::uuids::nil_uuid())
+        {
+            query.where.add(t.trade_id == boost::uuids::to_string(*trade_id_));
+        }
+        else
+        {
+            query.where.add(t.trade_id.is_null());
+        }
+    }
+
+    if (session_id_)
+    {
+        query.where.add(t.session_id == boost::uuids::to_string(*session_id_));
+    }
+
+    if (symbol_)
+    {
+        query.where.add(t.symbol == *symbol_);
+    }
+
+    if (exchange_name_)
+    {
+        query.where.add(t.exchange_name == enums::toString(*exchange_name_));
+    }
+
+    if (order_side_)
+    {
+        query.where.add(t.order_side == enums::toString(*order_side_));
+    }
+
+    if (order_type_)
+    {
+        query.where.add(t.order_type == enums::toString(*order_type_));
+    }
+
+    if (status_)
+    {
+        query.where.add(t.status == enums::toString(*status_));
+    }
+
+    if (created_at_)
+    {
+        query.where.add(t.created_at == *created_at_);
+    }
+}
+
+// Return a dictionary representation of the order
+nlohmann::json ct::db::Order::toJson() const
+{
+    nlohmann::json result;
+
+    result["id"]         = getIdAsString();
+    result["trade_id"]   = getTradeIdAsString();
+    result["session_id"] = getSessionIdAsString();
+
+    if (exchange_id_)
+    {
+        result["exchange_id"] = *exchange_id_;
+    }
+    else
+    {
+        result["exchange_id"] = nullptr;
+    }
+
+    result["symbol"]     = symbol_;
+    result["order_side"] = order_side_;
+    result["order_type"] = order_type_;
+    result["qty"]        = qty_;
+    result["filled_qty"] = filled_qty_;
+
+    if (price_)
+    {
+        result["price"] = *price_;
+    }
+    else
+    {
+        result["price"] = nullptr;
+    }
+
+    result["status"]     = status_;
+    result["created_at"] = created_at_;
+
+    if (canceled_at_)
+    {
+        result["canceled_at"] = *canceled_at_;
+    }
+    else
+    {
+        result["canceled_at"] = nullptr;
+    }
+
+    if (executed_at_)
+    {
+        result["executed_at"] = *executed_at_;
+    }
+    else
+    {
+        result["executed_at"] = nullptr;
+    }
+
+    return result;
+}
+
 // Default constructor
 ct::db::Candle::Candle() : id_(boost::uuids::random_generator()()) {}
 
@@ -284,6 +1629,202 @@ ct::db::Candle::Candle(const std::unordered_map< std::string, std::any >& attrib
     catch (const std::bad_any_cast& e)
     {
         throw std::runtime_error(std::string("Error initializing Candle: ") + e.what());
+    }
+}
+
+ct::db::Candle::Candle(
+
+    int64_t timestamp,
+    double open,
+    double close,
+    double high,
+    double low,
+    double volume,
+    enums::ExchangeName exchange_name,
+    std::string symbol,
+    enums::Timeframe timeframe)
+    : id_(boost::uuids::random_generator()())
+    , timestamp_(timestamp)
+    , open_(open)
+    , close_(close)
+    , high_(high)
+    , low_(low)
+    , volume_(volume)
+    , exchange_name_(exchange_name)
+    , symbol_(symbol)
+    , timeframe_(timeframe)
+{
+}
+
+// Convert DB row to model instance
+template < typename ROW >
+ct::db::Candle ct::db::Candle::fromRow(const ROW& row)
+{
+    Candle candle;
+    candle.id_            = boost::uuids::string_generator()(row.id.value());
+    candle.timestamp_     = row.timestamp;
+    candle.open_          = row.open;
+    candle.close_         = row.close;
+    candle.high_          = row.high;
+    candle.low_           = row.low;
+    candle.volume_        = row.volume;
+    candle.exchange_name_ = enums::toExchangeName(row.exchange_name);
+    candle.symbol_        = row.symbol;
+    candle.timeframe_     = enums::toTimeframe(row.timeframe);
+    return candle;
+}
+
+auto ct::db::Candle::prepareSelectStatementForConflictCheck(const CandlesTable& t,
+                                                            sqlpp::postgresql::connection& conn) const
+{
+    auto query  = dynamic_select(conn, all_of(t)).from(t).dynamic_where();
+    auto filter = Filter()
+                      .withExchangeName(exchange_name_)
+                      .withSymbol(symbol_)
+                      .withTimeframe(timeframe_)
+                      .withTimestamp(timestamp_);
+
+    filter.applyToQuery(query, t);
+
+    return query;
+}
+
+// Prepare insert statement
+auto ct::db::Candle::prepareInsertStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const
+{
+    return sqlpp::dynamic_insert_into(conn, t).dynamic_set(t.id            = getIdAsString(),
+                                                           t.timestamp     = timestamp_,
+                                                           t.open          = open_,
+                                                           t.close         = close_,
+                                                           t.high          = high_,
+                                                           t.low           = low_,
+                                                           t.volume        = volume_,
+                                                           t.exchange_name = enums::toString(exchange_name_),
+                                                           t.symbol        = symbol_,
+                                                           t.timeframe     = enums::toString(timeframe_));
+}
+
+// Prepare update statement
+auto ct::db::Candle::prepareUpdateStatement(const CandlesTable& t, sqlpp::postgresql::connection& conn) const
+{
+    return sqlpp::dynamic_update(conn, t)
+        .dynamic_set(t.timestamp     = timestamp_,
+                     t.open          = open_,
+                     t.close         = close_,
+                     t.high          = high_,
+                     t.low           = low_,
+                     t.volume        = volume_,
+                     t.exchange_name = enums::toString(exchange_name_),
+                     t.symbol        = symbol_,
+                     t.timeframe     = enums::toString(timeframe_))
+        .dynamic_where(t.id == parameter(t.id));
+}
+
+template < typename Query, typename Table >
+void ct::db::Candle::Filter::applyToQuery(Query& query, const Table& t) const
+{
+    if (id_)
+    {
+        query.where.add(t.id == boost::uuids::to_string(*id_));
+    }
+    if (timestamp_)
+    {
+        query.where.add(t.timestamp == *timestamp_);
+    }
+    if (open_)
+    {
+        query.where.add(t.open == *open_);
+    }
+    if (close_)
+    {
+        query.where.add(t.close == *close_);
+    }
+    if (high_)
+    {
+        query.where.add(t.high == *high_);
+    }
+    if (low_)
+    {
+        query.where.add(t.low == *low_);
+    }
+    if (volume_)
+    {
+        query.where.add(t.volume == *volume_);
+    }
+    if (exchange_name_)
+    {
+        query.where.add(t.exchange_name == enums::toString(*exchange_name_));
+    }
+    if (symbol_)
+    {
+        query.where.add(t.symbol == *symbol_);
+    }
+    if (timeframe_)
+    {
+        query.where.add(t.timeframe == enums::toString(*timeframe_));
+    }
+}
+
+/**
+ * @brief Store candles data into the database
+ *
+ * @param exchange_name Exchange name
+ * @param symbol Trading symbol
+ * @param timeframe Timeframe string
+ * @param candles Blaze matrix containing candle data
+ */
+void ct::db::saveCandles(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                         const enums::ExchangeName& exchange_name,
+                         const std::string& symbol,
+                         const enums::Timeframe& timeframe,
+                         const blaze::DynamicMatrix< double >& candles)
+{
+    // Make sure the number of candles is more than 0
+    if (candles.rows() == 0)
+    {
+        throw std::runtime_error("No candles to store for " + enums::toString(exchange_name) + "-" + symbol + "-" +
+                                 enums::toString(timeframe));
+    }
+
+    // Use the provided connection if available, otherwise get the default connection
+    auto& conn    = *(conn_ptr ? conn_ptr : Database::getInstance().getConnection());
+    const auto& t = Candle::table();
+
+    // Create state guard for this connection
+    ConnectionStateGuard stateGuard(conn);
+
+    auto stmt = sqlpp::insert_into(t).columns(
+        t.id, t.timestamp, t.open, t.close, t.high, t.low, t.volume, t.exchange_name, t.symbol, t.timeframe);
+
+    for (size_t i = 0; i < candles.rows(); ++i)
+    {
+        auto id = boost::uuids::to_string(boost::uuids::random_generator()());
+        stmt.values.add(t.id            = id,
+                        t.timestamp     = static_cast< int64_t >(candles(i, 0)),
+                        t.open          = candles(i, 1),
+                        t.close         = candles(i, 2),
+                        t.high          = candles(i, 3),
+                        t.low           = candles(i, 4),
+                        t.volume        = candles(i, 5),
+                        t.exchange_name = enums::toString(exchange_name),
+                        t.symbol        = symbol,
+                        t.timeframe     = enums::toString(timeframe));
+    }
+
+    try
+    {
+        conn(stmt);
+    }
+    catch (const std::exception& e)
+    {
+        std::ostringstream oss;
+        oss << "Error saving candles: " << e.what();
+        logger::LOG.error(oss.str());
+
+        // Mark the connection for reset
+        stateGuard.markForReset();
+
+        throw;
     }
 }
 
@@ -869,3 +2410,115 @@ ct::db::Trade::Trade(const std::unordered_map< std::string, std::any >& attribut
         throw std::runtime_error(std::string("Error initializing Trade: ") + e.what());
     }
 }
+
+// TODO: Move other model impl from header file here.
+
+template std::optional< ct::db::Candle > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                          const boost::uuids::uuid& id);
+
+template std::optional< ct::db::ClosedTrade > ct::db::findById(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const boost::uuids::uuid& id);
+
+template std::optional< ct::db::DailyBalance > ct::db::findById(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const boost::uuids::uuid& id);
+
+template std::optional< ct::db::ExchangeApiKeys > ct::db::findById(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Log > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                       const boost::uuids::uuid& id);
+
+template std::optional< ct::db::NotificationApiKeys > ct::db::findById(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Option > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                          const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Orderbook > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                             const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Ticker > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                          const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Trade > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                         const boost::uuids::uuid& id);
+
+template std::optional< ct::db::Order > ct::db::findById(std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                                                         const boost::uuids::uuid& id);
+
+template std::optional< std::vector< ct::db::Candle > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Candle::Filter& filter);
+
+template std::optional< std::vector< ct::db::ClosedTrade > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::ClosedTrade::Filter& filter);
+
+template std::optional< std::vector< ct::db::DailyBalance > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::DailyBalance::Filter& filter);
+
+template std::optional< std::vector< ct::db::ExchangeApiKeys > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::ExchangeApiKeys::Filter& filter);
+
+template std::optional< std::vector< ct::db::Log > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Log::Filter& filter);
+
+template std::optional< std::vector< ct::db::NotificationApiKeys > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::NotificationApiKeys::Filter& filter);
+
+template std::optional< std::vector< ct::db::Option > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Option::Filter& filter);
+
+template std::optional< std::vector< ct::db::Orderbook > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Orderbook::Filter& filter);
+
+template std::optional< std::vector< ct::db::Ticker > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Ticker::Filter& filter);
+
+template std::optional< std::vector< ct::db::Trade > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Trade::Filter& filter);
+
+template std::optional< std::vector< ct::db::Order > > ct::db::findByFilter(
+    std::shared_ptr< sqlpp::postgresql::connection > conn_ptr, const ct::db::Order::Filter& filter);
+
+template void ct::db::save(ct::db::Candle& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::ClosedTrade& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::DailyBalance& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::ExchangeApiKeys& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Log& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::NotificationApiKeys& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Option& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Orderbook& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Ticker& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Trade& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
+
+template void ct::db::save(ct::db::Order& model,
+                           std::shared_ptr< sqlpp::postgresql::connection > conn_ptr,
+                           const bool update_on_conflict);
