+++ tests/Exchange_tests.cpp
@@ -1,10 +1,8 @@
-#include <future>
 #include "DB.hpp"
 #include "Enum.hpp"
 #include "Exception.hpp"
 #include "Exchange.hpp"
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
+
 #include <gtest/gtest.h>
 
 // Helper to create a test order
