+++ tests/Orderbook_test.cpp
@@ -1,25 +1,4 @@
-#include <cmath>
-#include <cstdlib>
-#include <cstring>
-#include <dlfcn.h>
-#include <random>
-#include <stdexcept>
-
-#include <zlib.h>
-#include <boost/beast/core/detail/base64.hpp>
-#include <boost/multiprecision/cpp_dec_float.hpp>
-#include <boost/uuid/uuid.hpp>
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
-#include <date/date.h>
-#include <nlohmann/json.hpp>
-#include <openssl/sha.h>
-
 #include "Orderbook.hpp"
-
-#ifdef _WIN32
-#include <windows.h>
-#endif
 
 #include <gtest/gtest.h>
 
@@ -34,43 +13,43 @@ class OrderbooksStateTest : public ::testing::Test
 // Tests for orderbookTrimPrice
 TEST_F(OrderbooksStateTest, Basic)
 {
-    // Test ascending orderct::orderbook::OrderbooksState::trimPrice
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, true, 1.0), 100.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.1, true, 1.0), 101.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.9, true, 1.0), 101.0);
+    // Test ascending orderct::orderbook::OrderbooksState::trim
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, true, 1.0), 100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.1, true, 1.0), 101.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.9, true, 1.0), 101.0);
 
     // Test descending order
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, false, 1.0), 100.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.1, false, 1.0), 100.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.9, false, 1.0), 100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, false, 1.0), 100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.1, false, 1.0), 100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.9, false, 1.0), 100.0);
 }
 
 TEST_F(OrderbooksStateTest, EdgeCases)
 {
     // Test with very small unit
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.123456, true, 0.0001), 100.1235);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.123456, false, 0.0001), 100.1234);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.123456, true, 0.0001), 100.1235);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.123456, false, 0.0001), 100.1234);
 
     // Test with very large unit
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, true, 1000.0), 1000.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, false, 1000.0), 0.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, true, 1000.0), 1000.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, false, 1000.0), 0.0);
 
     // Test with unit equal to price
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, true, 100.0), 100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, true, 100.0), 100.0);
     // FIXME:
-    // EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(100.0, false, 100.0), 0.0);
+    // EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(100.0, false, 100.0), 0.0);
 
     // Test with zero price
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(0.0, true, 1.0), 0.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(0.0, false, 1.0), 0.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(0.0, true, 1.0), 0.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(0.0, false, 1.0), 0.0);
 
     // Test with negative price
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(-100.1, true, 1.0), -100.0);
-    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trimPrice(-100.1, false, 1.0), -101.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(-100.1, true, 1.0), -100.0);
+    EXPECT_DOUBLE_EQ(ct::orderbook::OrderbooksState::trim(-100.1, false, 1.0), -101.0);
 
     // Test invalid unit
-    EXPECT_THROW(ct::orderbook::OrderbooksState::trimPrice(100.0, true, 0.0), std::invalid_argument);
-    EXPECT_THROW(ct::orderbook::OrderbooksState::trimPrice(100.0, true, -1.0), std::invalid_argument);
+    EXPECT_THROW(ct::orderbook::OrderbooksState::trim(100.0, true, 0.0), std::invalid_argument);
+    EXPECT_THROW(ct::orderbook::OrderbooksState::trim(100.0, true, -1.0), std::invalid_argument);
 }
 
 // Stress tests
@@ -85,7 +64,7 @@ TEST_F(OrderbooksStateTest, Stress)
         double unit    = unit_dist(rng);
         bool ascending = (rng() % 2) == 0;
 
-        double result = ct::orderbook::OrderbooksState::trimPrice(price, ascending, unit);
+        double result = ct::orderbook::OrderbooksState::trim(price, ascending, unit);
         EXPECT_TRUE(std::isfinite(result));
         EXPECT_GE(result, 0.0);
     }
